### 2020-03-17 #2166 @whedon
[REVIEW]: COINSTAC: Collaborative Informatics and Neuroimaging Suite Toolkit for Anonymous Computation
**Submitting author:** @hvgazula (<a href="http://orcid.org/0000-0001-9318-8871">Harshvardhan Gazula</a>)
**Repository:** <a href="https://github.com/trendscenter/coinstac" target ="_blank">https://github.com/trendscenter/coinstac</a>
**Version:** v5.1.3

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/MRN-Code/coinstac">repository url</a>? (correction: moved and being redirected to https://github.com/trendscenter/coinstac now)
- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? (needs work: https://github.com/trendscenter/coinstac/issues/879, https://github.com/trendscenter/coinstac/issues/876)
