### 2020-03-20 trendscenter/coinstac#849 @gkiar
joss review comments/issues
Hi! I'm reviewing this library for JOSS (https://github.com/openjournals/joss-reviews/issues/2166), and will provide some feedback here as I'm going through the process of reviewing this package. I'll update this issue as I poke at this review the next few days, and please don't hesitate to make comments/responses in the meantime.

Note: these comments are based on the branch: `paper-branch`, and assumes that this will be merged directly to `master` upon completion of the review/requested changes.

- [x] Repository link in the submission is out of date (redirect from `MRN-Code` currently works, but the link should be corrected since a new one is available prior to publication)
- [x] The list of authors is inconsistent with the list of contributors for this project. Please clarify why none of the six top contributors to this project are authors, and why many of the listed authors are listed despite not having made clear contributions to the project according to the project's commit log.

- [x] After following the installation instructions [here](https://github.com/trendscenter/coinstac/blob/paper-branch/SETUP.md), it appears as though I was able to get to the point of launching `coinstac-ui`, but it hangs on a loading screen in the electron app after I launch `npm start`. There is nothing in the debug/console log to suggest why this is the case.
- [x] I cannot test the functionality of the app until I get past the point above

- [x] While the `npm` commands take care of installing dependencies, it would be nice to have the packages that are installed stated explicitly in the documentation.
- [x] While `SETUP.md` contains a link to a YouTube video showing how to use the tool, there is no accessible (text/figures only) version of these instructions for those with limited bandwidth, who I understand to be a target audience for this tool.
- [x] I was unable to find API-level documentation on how to use the functionality presented within the ui; while the installation did not appear to work as intended, stated above, I cannot really provide a more specific example as to what pieces are unclear or perhaps don't have integrated documentation
- [x] No clear automated testing/continuous integration setup
- [ ] Limited contribution guidelines for developers or those seeking support
  - No contributions markdown file or mention of how developers can become involved in the readme
  - Issue template is good for app crashes but limited for all other forms of support
  - Pull request template is incomplete

- [x] The paper makes no mention of other tools that have similar aims, such as CBRAIN, Loris, XNAT, OpenNeuro, BrainLife, or others.
- [x] Once the above-mentioned tools are explicitly stated and compared, they should be referenced accordingly.
