### 2020-03-20 #2166 @gkiar
@cMadan I have left my initial review at https://github.com/trendscenter/coinstac/issues/849 and checked off the already-met requirements in my form above.
