### 2020-03-23 4570493
Tweaks to get latest dpSVM working in UI
