### 2020-03-23 4cfe225
Change persistence strat to redux-persist
