+++ packages/coinstac-ui/app/render/index.js
@@ -5,6 +5,7 @@ import { hashHistory } from 'react-router';
 import { ipcRenderer, remote } from 'electron';
 import { ApolloProvider } from 'react-apollo';
 import { syncHistoryWithStore } from 'react-router-redux';
+import { PersistGate } from 'redux-persist/integration/react'
 import getApolloClient from './state/apollo-client';
 import configureStore from './state/store';
 import { start as startErrorHandling } from './utils/boot/configure-error-handling';
@@ -24,13 +25,15 @@ const rootEl = document.getElementById('app');
 global.config = remote.getGlobal('config');
 
 const client = getApolloClient(global.config);
-const store = configureStore(client);
+const { store, persistor } = configureStore(client);
 
-const history = syncHistoryWithStore(hashHistory, store)
+const history = syncHistoryWithStore(hashHistory, store);
 
 render(
   <ApolloProvider store={store} client={client}>
-    <Root history={history} />
+    <PersistGate loading={null} persistor={persistor}>
+      <Root history={history} />
+    </PersistGate>
   </ApolloProvider>,
   rootEl
 );
@@ -44,7 +47,9 @@ if (module.hot) {
     /* eslint-enable global-require */
     render(
       <ApolloProvider store={store} client={client}>
-        <NextRoot history={history} store={store} />
+        <PersistGate loading={null} persistor={persistor}>
+          <NextRoot history={history} store={store} />
+        </PersistGate>
       </ApolloProvider>,
       rootEl
     );
