### 2020-03-23 45ae4ed
Change persistence strat to redux-persist
