+++ packages/coinstac-ui/app/render/state/ducks/runs.js
@@ -1,75 +1,33 @@
 import { uniqBy } from 'lodash';
 import { applyAsyncLoading } from './loading';
-import { getDatabaseInstance } from '../local-db';
 
 // Actions
 const CLEAR_RUNS = 'CLEAR_RUNS';
-const GET_DB_RUNS = 'GET_DB_RUNS';
 const SAVE_LOCAL_RUN = 'SAVE_LOCAL_RUN';
-const SAVE_REMOTE_RUNS_LOCALLY = 'SAVE_REMOTE_RUNS_LOCALLY';
 const UPDATE_LOCAL_RUN = 'UPDATE_LOCAL_RUN';
 
 // Action Creators
-export const saveRemoteRunsLocally = applyAsyncLoading(
-  runs => dispatch => getDatabaseInstance().runs.bulkPut(runs)
-    .then(() => {
-      dispatch(({
-        type: SAVE_REMOTE_RUNS_LOCALLY,
-        payload: runs,
-      }));
-    })
-);
-
-export const clearRuns = applyAsyncLoading(() => dispatch => getDatabaseInstance().runs.clear()
-  .then(() => {
-    dispatch(({
-      type: CLEAR_RUNS,
-      payload: null,
-    }));
+export const clearRuns = applyAsyncLoading(() => (dispatch) => {
+  dispatch(({
+    type: CLEAR_RUNS,
+    payload: null,
   }));
+});
 
-export const getRunsForConsortium = applyAsyncLoading(consId => () => getDatabaseInstance().runs
-  .filter(run => run.consortiumId === consId)
-  .sortBy('startDate')
-  .then((runs) => {
-    return runs;
+export const saveLocalRun = applyAsyncLoading(run => (dispatch) => {
+  dispatch(({
+    type: SAVE_LOCAL_RUN,
+    payload: run,
   }));
-
-export const getLocalRun = applyAsyncLoading(runId => () => getDatabaseInstance().runs.get(runId)
-  .then((run) => {
-    return run;
-  }));
-
-export const getDBRuns = applyAsyncLoading(() => dispatch => getDatabaseInstance().runs
-  // .where('id').startsWith('local')
-  .toArray()
-  .then((runs) => {
-    dispatch(({
-      type: GET_DB_RUNS,
-      payload: runs,
-    }));
-    return runs;
-  }));
-
-export const saveLocalRun = applyAsyncLoading(run => dispatch => getDatabaseInstance().runs.put(run)
-  .then((key) => {
-    dispatch(({
-      type: SAVE_LOCAL_RUN,
-      payload: run,
-    }));
-    return key;
-  }));
+});
 
 export const updateLocalRun = applyAsyncLoading(
-  (runId, object) => dispatch => getDatabaseInstance().runs.update(runId, { ...object })
-    .then((key) => {
-      dispatch(({
-        type: UPDATE_LOCAL_RUN,
-        payload: { runId, object },
-      }));
-      return key;
-    })
-);
+  (runId, object) => (dispatch) => {
+    dispatch(({
+      type: UPDATE_LOCAL_RUN,
+      payload: { runId, object },
+    }));
+  });
 
 const INITIAL_STATE = {
   localRuns: [],
@@ -85,22 +43,6 @@ export default function reducer(state = INITIAL_STATE, action) {
   switch (action.type) {
     case CLEAR_RUNS:
       return INITIAL_STATE;
-    case GET_DB_RUNS: {
-      const localRuns = [];
-      const remoteRuns = [];
-
-      action.payload.forEach((run) => {
-        if (run.type === 'local') {
-          localRuns.push(run);
-        } else {
-          remoteRuns.push(run);
-        }
-      });
-
-      return {
[…]
