### 2020-03-24 1e1e531
Save api token in localstorage by user
