+++ packages/coinstac-ui/app/render/state/ducks/auth.js
@@ -8,6 +8,13 @@ import { applyAsyncLoading } from './loading';
 const apiServer = remote.getGlobal('config').get('apiServer');
 const API_URL = `${apiServer.protocol}//${apiServer.hostname}${apiServer.port ? `:${apiServer.port}` : ''}${apiServer.pathname}`;
 
+const API_TOKEN_KEY = 'id_token';
+let currentApiTokenKey = null;
+
+export function getCurrentApiTokenKey() {
+  return currentApiTokenKey;
+}
+
 const getErrorDetail = error => ({
   message: get(error, 'response.data.message'),
   statusCode: get(error, 'response.status'),
@@ -68,10 +75,12 @@ const initCoreAndSetToken = (reqUser, data, appDirectory, dispatch) => {
 
       ipcPromise.send('login-success', data.user.id);
 
+      currentApiTokenKey = `${API_TOKEN_KEY}_${data.user.id}`;
+
       if (reqUser.saveLogin) {
-        localStorage.setItem('id_token', data.id_token);
+        localStorage.setItem(getCurrentApiTokenKey(), data.id_token);
       } else {
-        sessionStorage.setItem('id_token', data.id_token);
+        sessionStorage.setItem(getCurrentApiTokenKey(), data.id_token);
       }
 
       dispatch(setUser(user));
@@ -79,20 +88,21 @@ const initCoreAndSetToken = (reqUser, data, appDirectory, dispatch) => {
 };
 
 export const logout = applyAsyncLoading(() => (dispatch) => {
-  localStorage.removeItem('id_token');
-  sessionStorage.removeItem('id_token');
+  localStorage.removeItem(getCurrentApiTokenKey());
+  sessionStorage.removeItem(getCurrentApiTokenKey());
   return ipcPromise.send('logout')
     .then(() => {
       dispatch(clearUser());
+      currentApiTokenKey = null;
     });
 });
 
 export const autoLogin = applyAsyncLoading(() => (dispatch, getState) => {
-  let token = localStorage.getItem('id_token');
+  let token = localStorage.getItem(getCurrentApiTokenKey());
   let saveLogin = true;
 
   if (!token || token === 'null' || token === 'undefined') {
-    token = sessionStorage.getItem('id_token');
+    token = sessionStorage.getItem(getCurrentApiTokenKey());
     saveLogin = false;
   }
 
