### 2020-03-24 05176f8
Fix bug on startup
