### 2020-03-25 f55de6e
Fix state rehydrate on consecutive logins
