+++ packages/coinstac-ui/app/render/state/ducks/app.js
@@ -3,22 +3,14 @@ const INITIAL_STATE = {
 };
 
 // Actions
-const CLEAR_LOGS = 'CLEAR_LOGS';
 const APPEND_LOG_MESSAGE = 'APPEND_LOG_MESSAGE';
 
 // Action Creators
-export const clearLogs = () => ({ type: CLEAR_LOGS });
-
 export const appendLogMessage = message => ({ type: APPEND_LOG_MESSAGE, payload: message });
 
 // Reducer
 export default function reducer(state = INITIAL_STATE, action) {
   switch (action.type) {
-    case CLEAR_LOGS:
-      return {
-        ...state,
-        logs: null,
-      };
     case APPEND_LOG_MESSAGE:
       return {
         ...state,
