+++ packages/coinstac-ui/app/main/index.js
@@ -157,6 +157,8 @@ loadConfig()
      * @type {[type]}
      */
     ipcPromise.on('logout', () => {
+      mainWindow.webContents.send('logout');
+
       // TODO: hacky way to not get a mqtt reconnn loop
       // a better way would be to make an actual shutdown fn for pipeline
       return new Promise((resolve) => {
