### 2020-03-28 trendscenter/coinstac#849 @hvgazula
> Hi! I'm reviewing this library for JOSS ([openjournals/joss-reviews#2166](https://github.com/openjournals/joss-reviews/issues/2166)), and will provide some feedback here as I'm going through the process of reviewing this package. I'll update this issue as I poke at this review the next few days, and please don't hesitate to make comments/responses in the meantime.
>
> Note: these comments are based on the branch: `paper-branch`, and assumes that this will be merged directly to `master` upon completion of the review/requested changes.
>
> ### General
> * [ ]  Repository link in the submission is out of date (redirect from `MRN-Code` currently works, but the link should be corrected since a new one is available prior to publication)
Thanks for pointing this out. I am not sure how to change the repository link in the submission. Please advise.
> * [ ]  The list of authors is inconsistent with the list of contributors for this project. Please clarify why none of the six top contributors to this project are authors, and why many of the listed authors are listed despite not having made clear contributions to the project according to the project's commit log.
Some of the top contributors (from github list) are no longer associated with the project and some are functioning as paid consultants on the project, in addition computation author contributions are in multiple other repositories. We have reworked the author list to be more comprehensive, we now include everyone who is actively participating in either the platform or computational code development or management and oversight of the project. We have also added a list of those who left the COINSTAC team prior to the start of this submission to the acknowledgement section for completeness.

>
> ### Functionality
> * [ ]  After following the installation instructions [here](https://github.com/trendscenter/coinstac/blob/paper-branch/SETUP.md), it appears as though I was able to get to the point of launching `coinstac-ui`, but it hangs on a loading screen in the electron app after I launch `npm start`. There is nothing in the debug/console log to suggest why this is the case.
Those instructions are for developing and contributing to the coinstac codebase. For end user instructions please check [here](https://github.com/trendscenter/coinstac-instructions). A link to these instructions has also been added to the main project README.

> * [ ]  I cannot test the functionality of the app until I get past the point above
See above.
>
> ### Documentation
> * [ ]  While the `npm` commands take care of installing dependencies, it would be nice to have the packages that are installed stated explicitly in the documentation.
The packages are listed [here](https://github.com/trendscenter/coinstac/blob/master/package.json)
> * [ ]  While `SETUP.md` contains a link to a YouTube video showing how to use the tool, there is no accessible (text/figures only) version of these instructions for those with limited bandwidth, who I understand to be a target audience for this tool.
The instruction are [here](https://github.com/trendscenter/coinstac-instructions)
> * [ ]  I was unable to find API-level documentation on how to use the functionality presented within the ui; while the installation did not appear to work as intended, stated above, I cannot really provide a more specific example as to what pieces are unclear or perhaps don't have integrated documentation
Please see instruction [here](https://github.com/trendscenter/coinstac-instructions/blob/master/How%20to%20Run%20an%20Analysis%20on%20COINSTAC.docx)
> * [ ]  No clear automated testing/continuous integration setup
End to end integration tests can be found in the packages/coinstac-ui. We are currently in the process for rewriting necessary dependencies for use with CircleCI.
> * [ ]  Limited contribution guidelines for developers or those seeking support
>
>   * No contributions markdown file or mention of how developers can become involved in the readme
>   * Issue template is good for app crashes but limited for all other forms of support
>   * Pull request template is incomplete
We should make a contributions template. App errors have been the most challenging for end users to submit so the template leans that way, but could be expanded. Unsure how the PR template is incomplete but open to suggestions.
>
> ### Software Paper
> * [ ]  The paper makes no mention of other tools that have similar aims, such as CBRAIN, Loris, XNAT, OpenNeuro, BrainLife, or others.
The aforementioned tools although are open-source informatics platforms dedicated to imaging-based research none of them are decentralized platforms. COINSTAC is rather a decentralized, privacy preserving analysis tool, and as such it is quite distinct in this regard from the tools above and also from even the COINS platform which is a centralized repository tool. which is exactly what COINSTAC is. We have also added references to each of the above tools, but do not do a more detailed comparison as COINSTAC is, to our knowledge, the first of its kind application platform enabling decentralized analysis of brain imaging data.

> * [ ]  Once the above-mentioned tools are explicitly stated and compared, they should be referenced accordingly.
COINSTAC is, to our knowledge, the first of its kind application platform enabling decentralized analysis of brain imaging data.
