### 2020-03-31 4977e1a
Updated user model to save app roles
