### 2020-03-31 7b37e1d
Added role type to updating user roles resolvers
