+++ packages/coinstac-ui/app/render/state/graphql/props.js
@@ -165,9 +165,9 @@ export const saveDocumentProp = (funcName, objVar) => {
 export const userRolesProp = (name) => {
   return {
     props: ({ mutate }) => ({
-      [name]: (userId, table, doc, role) => mutate({
+      [name]: (userId, table, doc, role, roleType) => mutate({
         variables: {
-          userId, table, doc, role,
+          userId, table, doc, role, roleType,
         },
       }),
     }),
