+++ packages/coinstac-ui/app/render/utils/helpers.js
@@ -1,6 +1,9 @@
-import { indexOf } from 'lodash';
+import { get, indexOf } from 'lodash';
 
-// eslint-disable-next-line import/prefer-default-export
 export const isPipelineOwner = (permissions, owningConsortium) => {
   return indexOf(permissions.consortia[owningConsortium], 'owner') !== -1;
+};
+
+export const isAdmin = (user) => {
+  return get(user, 'permissions.roles.admin', false);
 };
