### 2020-03-31 139c97f
Logout logged in user when his admin role is removed
