### 2020-04-01 2f82f77
Updated computation delete
