### 2020-04-02 5c167f5
PR review changes, deletions, and tweaks
