### 2020-04-02 c59927b
PR review changes, deletions, and tweaks
