+++ packages/coinstac-api-server/src/data/schema.js
@@ -155,7 +155,8 @@ const typeDefs = `
   # This is the general mutation description
   type Mutation {
     # Stringify incoming computation, parse prior to insertion call
-    addComputation(computationSchema: ComputationInput): Computation
+    createComputation(computationSchema: ComputationInput): Computation
+    updateComputation(computationId: ID!, computationSchema: ComputationInput): Computation
     addUserRole(userId: ID!, table: String!, doc: String!, role: String!, roleType: String!): User
     createRun(consortiumId: ID): Run
     deleteConsortiumById(consortiumId: ID): Consortium
