+++ packages/coinstac-ui/app/render/state/graphql/props.js
@@ -162,6 +162,16 @@ export const saveDocumentProp = (funcName, objVar) => {
   };
 };
 
+export const genericProp = (funcName) => {
+  return {
+    props: ({ mutate }) => ({
+      [funcName]: obj => mutate({
+        variables: obj,
+      }),
+    }),
+  };
+};
+
 export const userRolesProp = (name) => {
   return {
     props: ({ mutate }) => ({
