### 2020-04-09 1f89a31
Added command for computation upload in simulator
