### 2020-04-09 9636227
Connected commander arguments to computation upload simulator
