+++ packages/coinstac-api-server/src/routes/auth.js
@@ -10,9 +10,15 @@ module.exports = [
         { method: helperFunctions.validateUser, assign: 'user' },
       ],
       handler: (req, res) => {
+        const user = {
+          ...req.pre.user,
+        };
+
+        delete user.passwordHash;
+
         res({
-          id_token: helperFunctions.createToken(req.pre.user.id),
-          user: req.pre.user,
+          id_token: helperFunctions.createToken(user.username),
+          user,
         }).code(201);
       },
     },
@@ -25,14 +31,14 @@ module.exports = [
       handler: ({
         auth: {
           credentials: {
-            email, id, institution, permissions,
+            email, username, institution, permissions,
           },
         },
       }, res) => {
         res({
-          id_token: helperFunctions.createToken(id),
+          id_token: helperFunctions.createToken(username),
           user: {
-            email, id, institution, permissions,
+            email, username, institution, permissions,
           },
         }).code(201);
       },
@@ -50,12 +56,12 @@ module.exports = [
         helperFunctions.hashPassword(req.payload.password)
           .then(passwordHash => helperFunctions.createUser(req.payload, passwordHash))
           .then(({
-            id, institution, email, permissions,
+            _id, username, institution, email, permissions,
           }) => {
             res({
-              id_token: helperFunctions.createToken(id),
+              id_token: helperFunctions.createToken(username),
               user: {
-                id, institution, email, permissions,
+                id: _id, username, institution, email, permissions,
               },
             }).code(201);
           });
