+++ packages/coinstac-api-server/src/data/resolvers.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-await-in-loop */
 const rethink = require('rethinkdb');
 const Boom = require('boom');
 const GraphQLJSON = require('graphql-type-json');
@@ -8,149 +9,138 @@ const { uniq } = require('lodash');
 const helperFunctions = require('../auth-helpers');
 const initSubscriptions = require('./subscriptions');
 const config = require('../../config/default');
+const database = require('../database');
 
-/**
- * Helper function to retrieve all members of given table
- * @param {string} table - The table name
- * @return {array} The contents of the requested table
- */
-function fetchAll(table) {
-  let connection;
-  return helperFunctions.getRethinkConnection()
-    .then((db) => {
-      connection = db;
-      return rethink.table(table).orderBy({ index: 'id' }).run(connection);
-    })
-    .then(cursor => cursor.toArray())
-    .then(res => connection.close().then(() => res));
-}
+async function fetchOnePipeline(id) {
+  const db = database.getDbInstance();
 
-/**
- * Helper function to retrieve a single entry in a table
- * @param {string} table - The table name
- * @param {string} id - The entry id
- * @return {object} The requested table entry
- */
-function fetchOne(table, id) {
-  return helperFunctions.getRethinkConnection()
-    .then(connection => rethink.table(table).get(id).run(connection));
-}
+  const pipelineSteps = await db.collection('pipelines').aggregate([
+    { $match: { _id: id } },
+    { $unwind: '$steps' },
+    {
+      $lookup: {
+        from: 'computations',
+        localField: 'steps.computations',
+        foreignField: '_id',
+        as: 'steps.computations',
+      },
+    },
+  ]);
 
-function fetchOnePipeline(id) {
-  return helperFunctions.getRethinkConnection()
-    .then(connection => rethink.table('pipelines')
-      .get(id)
-    // Populate computations subfield with computation meta information
-      .merge(pipeline => ({
-        steps: pipeline('steps').map(step => step.merge({
-          computations: step('computations').map(compId => rethink.table('computations').get(compId)),
-        })),
-      }))
-      .run(connection)
-      .then(res => connection.close().then(() => res)))
-    .then(result => result);
+  if (!pipelineSteps || pipelineSteps.length === 0) {
+    return null;
+  }
+
+  let pipe = null;
+  while (await pipelineSteps.hasNext()) {
+    const currentStep = await pipelineSteps.next();
+
+    if (!pipe) {
+      pipe = {
+        ...currentStep,
+        steps: [],
+      };
+    }
+
+    pipe.steps.push(currentStep.steps);
+  }
+
+  return pipe;
 }
 
 /**
  * Helper function for add permissions to an user
- * @param {object} connection - Existing db connection
  * @param {object} args - Update object
  * @param {string} args.userId - Id of the user which will have permissions changed
  * @param {string} args.role - Role of the user
  * @param {string} args.doc - Id of the document for which the user will gain access
  * @param {string} args.table - Table of the document for which the user will gain access
  */
-async function addUserPermissions(connection, args) {
-  const perms = await rethink.table('users').get(args.userId)('permissions').run(connection);
+async function addUserPermissions(args) {
+  const db = database.getDbInstance();
 
   const { role, doc, table } = args;
 
-  let newRoles = [role];
   const promises = [];
 
-  // Grab existing roles if present
-  if (perms[table][doc] && perms[table][doc].indexOf(role) === -1) {
-    newRoles = newRoles.concat(perms[table][doc]);
-  } else if (perms[table][doc]) {
-    newRoles = perms[table][doc];
-  }
+  const updateObj = {
+    $addToSet: {
+      [`permissions.${table}.${doc}`]: role,
+    },
+  };
 
[…]
