+++ packages/coinstac-api-server/src/auth-helpers.js
@@ -4,6 +4,7 @@ const jwt = require('jsonwebtoken');
 const rethink = require('rethinkdb');
 const Promise = require('bluebird');
 const config = require('../config/default');
+const database = require('./database');
 
 let dbmap;
 try {
@@ -38,35 +39,37 @@ const helperFunctions = {
    * @param {string} passwordHash string of hashed password
    * @return {object} The updated user object
    */
-  createUser(user, passwordHash) {
-    return helperFunctions.getRethinkConnection()
-      .then((connection) => {
-        const userDetails = {
-          id: user.username,
-          email: user.email,
-          institution: user.institution,
-          passwordHash,
-          permissions: {
-            computations: {},
-            consortia: {},
-            pipelines: {},
-          },
-          consortiaStatuses: {},
-        };
+  async createUser(user, passwordHash) {
+    const userDetails = {
+      username: user.username,
+      email: user.email,
+      institution: user.institution,
+      passwordHash,
+      permissions: {
+        computations: {},
+        consortia: {},
+        pipelines: {},
+      },
+      consortiaStatuses: {},
+    };
 
-        if (user.permissions) {
-          userDetails.permissions = user.permissions;
-        }
+    if (user._id) {
+      userDetails._id = user._id;
+    }
 
-        if (user.consortiaStatuses) {
-          userDetails.consortiaStatuses = user.consortiaStatuses;
-        }
+    if (user.permissions) {
+      userDetails.permissions = user.permissions;
+    }
 
-        return rethink.table('users')
-          .insert(userDetails, { returnChanges: true })
-          .run(connection).then(res => connection.close().then(() => res));
-      })
-      .then(result => result.changes[0].new_val);
+    if (user.consortiaStatuses) {
+      userDetails.consortiaStatuses = user.consortiaStatuses;
+    }
+
+    const db = database.getDbInstance();
+
+    const result = await db.collection('users').insertOne(userDetails);
+
+    return result.ops[0];
   },
   /**
    * dbmap getter
@@ -91,17 +94,13 @@ const helperFunctions = {
   },
   /**
    * Returns user table object for requested user
-   * @param {object} credentials credentials of requested user
+   * @param {string} username username of requested user
    * @return {object} The requested user object
    */
-  async getUserDetails(credentials) {
-    const connection = await helperFunctions.getRethinkConnection();
+  async getUserDetails(username) {
+    const db = database.getDbInstance();
 
-    const user = await rethink.table('users').get(credentials.username).run(connection);
-
-    await connection.close();
-
-    return user;
+    return db.collection('users').findOne({ username });
   },
   /**
    * Hashes password for storage in database
@@ -148,9 +147,9 @@ const helperFunctions = {
    * @param {function} callback function signature (err, isValid, alternative credentials)
    */
   validateToken(decoded, request, callback) {
-    helperFunctions.getUserDetails({ username: decoded.username })
+    helperFunctions.getUserDetails(decoded.username)
       .then((user) => {
-        if (user.id) {
+        if (user) {
           callback(null, true, user);
         } else {
           callback(null, false, null);
@@ -160,15 +159,16 @@ const helperFunctions = {
   /**
    * Confirms that submitted email is new
    * @param {object} req request
-   * @param {object} connection GraphQL connection
    * @return {boolean} Is the email unique?
    */
-  validateUniqueEmail(req, connection) {
-    return rethink.table('users')
-      .filter({ email: req.payload.email })
-      .count()
-      .eq(0)
-      .run(connection);
+  async validateUniqueEmail(req) {
+    const db = database.getDbInstance();
+
+    const count = await db.collection('users')
[…]
