+++ packages/coinstac-api-server/test/test-setup-old.js
@@ -1,0 +1,942 @@
+require('trace');
+require('clarify');
+const rethink = require('rethinkdb');
+const database = require('../src/database');
+
+const drneVbm = require('./data/coinstac-schema-regression-vbm');
+// const ssrVbm = require('./data/coinstac-schema-regression-ss-vbm');
+// const msrVbm = require('./data/coinstac-schema-regression-ms-vbm');
+
+const drneFsl = require('./data/coinstac-schema-regression-fsl');
+// const ssrFsl = require('./data/coinstac-schema-regression-ss-fsl');
+const msrFsl = require('./data/coinstac-schema-regression-ms-fsl');
+
+const gica = require('./data/coinstac-gica-pipeline');
+const ddfnc = require('./data/coinstac-ddfnc-pipeline');
+const dpsvm = require('./data/coinstac-dpsvm');
+const vbm = require('./data/coinstac-vbm-pre');
+
+const fmri = require('./data/coinstac-fmri');
+
+const decentralized = require('./data/coinstac-decentralized-test');
+const decentralizedError = require('./data/coinstac-decentralized-error');
+const enigmaSans = require('./data/coinstac-enigma-sans');
+const local = require('./data/coinstac-local-test');
+const localError = require('./data/coinstac-local-error');
+
+const helperFunctions = require('../src/auth-helpers');
+
+function populateComputations() {
+  const { db } = database;
+
+  const x = db.collection('computations').insertMany([
+    { ...local, submittedBy: 'author' },
+    { ...decentralized, submittedBy: 'author' },
+    { ...msrFsl, submittedBy: 'author' },
+    { ...vbm, submittedBy: 'author' },
+    { ...gica, submittedBy: 'author' },
+    { ...ddfnc, submittedBy: 'author' },
+    { ...dpsvm, submittedBy: 'author' },
+    { ...drneVbm, submittedBy: 'author' },
+    { ...drneFsl, submittedBy: 'author' },
+    { ...decentralizedError, submittedBy: 'author' },
+    { ...enigmaSans, submittedBy: 'author' },
+    { ...localError, submittedBy: 'author' },
+    { ...fmri, submittedBy: 'author' },
+  ]);
+
+  console.log(x);
+}
+
+async function populate() {
+  await populateComputations();
+}
+
+database.connect()
+  .then(() => {
+    database.dropDbInstance();
+    database.createDbInstance();
+
+    return populate();
+  });
+
+let connection;
+
+helperFunctions.getRethinkConnection()
+  .then((db) => {
+    connection = db;
+    return rethink.dbList().contains('coinstac')
+      .do((exists) => {
+        return rethink.branch(
+          exists,
+          rethink.dbDrop('coinstac'),
+          { dbs_dropped: 0 }
+        );
+      }).run(connection)
+      .then(() => {
+        if (process.argv[2]) {
+          return rethink.db('rethinkdb').table('users')
+            .get('admin').update({ password: process.argv[2] })
+            .run(connection)
+            .then(() => connection.close())
+            .then(() => {
+              helperFunctions.setDBMap({ rethinkdbAdmin: { user: 'admin', password: process.argv[2] } });
+              return helperFunctions.getRethinkConnection();
+            })
+            .then((newDBConn) => { connection = newDBConn; });
+        }
+      })
+      .then(() => rethink.dbCreate('coinstac').run(connection))
+      .then(() => rethink.tableCreate('pipelines').run(connection))
+      .then(() => rethink.tableCreate('computations').run(connection))
+      .then(() => rethink.table('computations').insert([
+        Object.assign({}, local, { submittedBy: 'author' }),
[…]
