### 2020-04-14 e679a41
add owner only result filtering
