+++ packages/coinstac-pipeline/src/pipeline-manager.js
@@ -223,9 +223,18 @@ module.exports = {
     };
 
 
-    const clientPublish = (clientList, data) => {
+    const clientPublish = (clientList, data, opts) => {
       clientList.forEach((client) => {
-        serverMqt.publish(`${client}-run`, JSON.stringify(data), { qos: 1 }, (err) => { if (err) logger.error(`Mqtt error: ${err}`); });
+        if (opts.success && opts.limitOutputToOwner) {
+          if (client === opts.owner) {
+            serverMqt.publish(`${client}-run`, JSON.stringify(data), { qos: 1 }, (err) => { if (err) logger.error(`Mqtt error: ${err}`); });
+          } else {
+            const limitedData = Object.assign({}, data, { files: undefined, output: { success: true, output: { message: 'output sent to consortium owner' } } });
+            serverMqt.publish(`${client}-run`, JSON.stringify(limitedData), { qos: 1 }, (err) => { if (err) logger.error(`Mqtt error: ${err}`); });
+          }
+        } else {
+          serverMqt.publish(`${client}-run`, JSON.stringify(data), { qos: 1 }, (err) => { if (err) logger.error(`Mqtt error: ${err}`); });
+        }
       });
     };
 
@@ -466,11 +475,16 @@ module.exports = {
             }
             break;
           case 'finished':
+          debugger
             if (activePipelines[data.runId] && activePipelines[data.runId].clients[data.id]) {
               if (activePipelines[data.runId].finalTransferList) {
                 activePipelines[data.runId].finalTransferList.add(data.id);
                 if (activePipelines[data.runId].clients
                   .every(value => activePipelines[data.runId].finalTransferList.has(value))
+                  || (
+                    activePipelines[data.runId].limitOutputToOwner
+                    && data.id === activePipelines[data.runId].owner
+                  )
                 ) {
                   cleanupPipeline(data.runId)
                     .catch(e => logger.error(`Pipeline cleanup failure: ${e}`));
@@ -636,6 +650,8 @@ module.exports = {
               operatingDirectory,
               clientId,
               userDirectories,
+              // owner: spec.owner,
+              owner: 'local0',
             }),
             baseDirectory: path.resolve(operatingDirectory, 'input', clientId, runId),
             cacheDirectory: userDirectories.cacheDirectory,
@@ -648,6 +664,10 @@ module.exports = {
             communicate: undefined,
             clients,
             remote: { reject: () => {} }, // noop for pre pipe errors
+            // owner: spec.owner,
+            // limitOutputToOwner: spec.limitOutputToOwner,
+            owner: 'local0',
+            limitOutputToOwner: true,
           },
           activePipelines[runId]
         );
@@ -750,13 +770,28 @@ module.exports = {
                           output: message,
                           files: [archiveName],
                           iteration: messageIteration,
+                        },
+                        {
+                          success: message.success,
+                          limitOutputToOwner: activePipelines[pipeline.id].limitOutputToOwner,
+                          owner: activePipelines[pipeline.id].owner,
                         }
                       );
                     });
                   }
                   clientPublish(
                     activePipelines[pipeline.id].clients,
-                    { runId: pipeline.id, output: message, iteration: messageIteration }
+                    {
+                      runId: pipeline.id,
+                      output: message,
+                      iteration: messageIteration,
+                    },
+                    {
+                      success: message.success,
+                      limitOutputToOwner: activePipelines[pipeline.id].limitOutputToOwner,
+                      owner: activePipelines[pipeline.id].owner,
+                    }
+
                   );
                 });
             }
@@ -947,9 +982,14 @@ module.exports = {
                 return res;
               });
           }).then((res) => {
+            debugger
             if (!activePipelines[runId].finalTransferList
[…]
