### 2020-04-14 5daf0e3
add owner only result filtering
