### 2020-04-16 fd21adc
Pipeline error out when pipeline server is down
