+++ packages/coinstac-api-server/README.md
@@ -2,14 +2,13 @@
 
 <img src="https://raw.githubusercontent.com/MRN-Code/coinstac/master/img/coinstac.png" height="75px">
 
-The API Server hosts the GraphQL endpoints called by the UI and other services accessing RethinkDB. 
+The API Server hosts the GraphQL endpoints called by the UI and other servicesB.
 
 ## Installing/Running
 
-  * Install RethinkDB: `brew update && brew install rethinkdb`
-  * Start RethinkDB: `rethinkdb`
+  * Start MongoDB: `docker-compose up -d`
   * Navigate into `packages/coinstac-api-server` and `npm i`
-  * Setup Rethink with test data: `npm run test-setup`
+  * Setup database with test data: `npm run test-setup`
   * Start server: `npm run start`
 
 ## License
