+++ packages/coinstac-api-server/test/test-setup.js
@@ -24,23 +24,51 @@ const enigmaSans = require('./data/coinstac-enigma-sans');
 const local = require('./data/coinstac-local-test');
 const localError = require('./data/coinstac-local-error');
 
+const CONSORTIA_IDS = [
+  database.createUniqueId(),
+  database.createUniqueId(),
+];
+
+const PIPELINE_IDS = [
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+];
+
+const COMPUTATION_IDS = [
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+  database.createUniqueId(),
+];
+
 async function populateComputations() {
   const db = database.getDbInstance();
 
   return db.collection('computations').insertMany([
-    { ...local, submittedBy: 'author' },
-    { ...decentralized, submittedBy: 'author' },
-    { ...msrFsl, submittedBy: 'author' },
-    { ...vbm, submittedBy: 'author' },
-    { ...gica, submittedBy: 'author' },
-    { ...ddfnc, submittedBy: 'author' },
-    { ...dpsvm, submittedBy: 'author' },
-    { ...drneVbm, submittedBy: 'author' },
-    { ...drneFsl, submittedBy: 'author' },
-    { ...decentralizedError, submittedBy: 'author' },
-    { ...enigmaSans, submittedBy: 'author' },
-    { ...localError, submittedBy: 'author' },
-    { ...fmri, submittedBy: 'author' },
+    { ...local, submittedBy: 'author', _id: COMPUTATION_IDS[0] },
+    { ...decentralized, submittedBy: 'author', _id: COMPUTATION_IDS[1] },
+    { ...msrFsl, submittedBy: 'author', _id: COMPUTATION_IDS[2] },
+    { ...vbm, submittedBy: 'author', _id: COMPUTATION_IDS[3] },
+    { ...gica, submittedBy: 'author', _id: COMPUTATION_IDS[4] },
+    { ...ddfnc, submittedBy: 'author', _id: COMPUTATION_IDS[5] },
+    { ...dpsvm, submittedBy: 'author', _id: COMPUTATION_IDS[6] },
+    { ...drneVbm, submittedBy: 'author', _id: COMPUTATION_IDS[7] },
+    { ...drneFsl, submittedBy: 'author', _id: COMPUTATION_IDS[8] },
+    { ...decentralizedError, submittedBy: 'author', _id: COMPUTATION_IDS[9] },
+    { ...enigmaSans, submittedBy: 'author', _id: COMPUTATION_IDS[10] },
+    { ...localError, submittedBy: 'author', _id: COMPUTATION_IDS[11] },
+    { ...fmri, submittedBy: 'author', _id: COMPUTATION_IDS[12] },
   ]);
 }
 
@@ -49,6 +77,7 @@ async function populateConsortia() {
 
   return db.collection('consortia').insertMany([
     {
+      _id: CONSORTIA_IDS[0],
       name: 'Test Consortia 1',
       description: 'This consortia is for testing.',
       owners: ['author'],
@@ -57,7 +86,8 @@ async function populateConsortia() {
       createDate: 1551333489519,
     },
     {
-      activePipelineId: 'test-pipeline-ssr',
+      _id: CONSORTIA_IDS[1],
+      activePipelineId: PIPELINE_IDS[0],
[…]
