### 2020-04-20 636bfaf
Remove persistence of local runs
