### 2020-04-21 4d660f6
Reverting function "getSubPathsAndGroupExtension" to working version
