+++ packages/coinstac-client-core/src/index.js
@@ -9,6 +9,8 @@ const fs = require('fs');
 const unlinkAsync = pify(fs.unlink);
 const linkAsync = pify(fs.link);
 const statAsync = pify(fs.stat);
+const readdirAsync = pify(fs.readdir);
+
 
 const path = require('path');
 const winston = require('winston');
@@ -182,17 +184,17 @@ class CoinstacClient {
     }
 
     // Iterate through all paths
-    group.paths.forEach(async (path) => {
-      let p = path;
+    await Promise.all(group.paths.map(async (filePath) => {
+      let p = filePath;
       // Combine path with parent dir to get absolute path
       if (group.parentDir) {
-        p = group.parentDir.concat(`/${p}`);
+        p = path.join(group.parentDir, p);
       }
 
-      const stats = await fs.statAsync(p);
+      const stats = await statAsync(p);
 
       if (stats.isDirectory()) {
-        const dirs = await fs.readdir(p);
+        const dirs = await readdirAsync(p);
         const paths = [...dirs.filter(item => !(/(^|\/)\.[^/.]/g).test(item))];
         // Recursively retrieve path contents of directory
         const subGroup = await this.getSubPathsAndGroupExtension({
@@ -200,7 +202,6 @@ class CoinstacClient {
           extension: group.extension,
           parentDir: p,
         });
-
         if (subGroup) {
           if (subGroup.error) {
             return subGroup;
@@ -224,8 +225,7 @@ class CoinstacClient {
         extension = thisExtension;
         pathsArray.push(p);
       }
-    });
-
+    }));
     return { paths: pathsArray, extension };
   }
 
