+++ packages/coinstac-ui/app/main/index.js
@@ -335,7 +335,6 @@ loadConfig()
         const { filesArray, steps } = runPipelineFunctions.parsePipelineInput(
           pipelineRun.pipelineSnapshot, dataMappings
         );
-
         const run = {
           ...pipelineRun,
           pipelineSnapshot: {
@@ -501,7 +500,7 @@ loadConfig()
       } else if (org === 'directory') {
         properties = ['openDirectory'];
         postDialogFunc = ipcFunctions.manualDirectorySelection;
-      } else if (org === 'bundle') {
+      } else if (org === 'bundle' || org === 'singles') {
         filters = [
           {
             name: 'File Types',
