### 2020-04-23 4cad83a
Modifications post merge with master
