### 2020-04-27 4bb949e
Updated SETUP.md to include link to dependencies
This is for JOSS reviewer.
