### 2020-04-27 trendscenter/coinstac#849 @cMadan
Just to follow-up re: authorship, the intention isn't for JOSS to be overly prescriptive and we completely understand that some individuals should be included as authors even though they haven't 'commit'-ed to the code directly. I think the authors here have been reasonable in including/acknowledging contributors, and the public discussion of this (as an issue in the repo) has also given an opportunity for someone to indicate that they may have disagreed with their recognition (if they felt that was the case).
