+++ packages/coinstac-api-server/README.md
@@ -2,7 +2,7 @@
 
 <img src="https://raw.githubusercontent.com/MRN-Code/coinstac/master/img/coinstac.png" height="75px">
 
-The API Server hosts the GraphQL endpoints called by the UI and other servicesB.
+The API Server hosts the GraphQL endpoints called by the UI and other services.
 
 ## Installing/Running
 
