### 2020-04-29 6bbaf4c
Store local run results only locally
