+++ packages/coinstac-ui/app/main/index.js
@@ -148,10 +148,6 @@ loadConfig()
           });
     });
 
-    ipcMain.on('login-success', (userId) => {
-      mainWindow.webContents.send('login-success', userId);
-    });
-
     /**
      * [initializedCore description]
      * @type {[type]}
