### 2020-04-29 611da60
Store local run results only locally
