+++ packages/coinstac-ui/app/render/state/ducks/auth.js
@@ -64,27 +64,28 @@ export const setApiVersionCheck = isApiVersionCompatible => ({
 });
 
 // Helpers
-const initCoreAndSetToken = (reqUser, data, appDirectory, dispatch) => {
+const initCoreAndSetToken = async (reqUser, data, appDirectory, dispatch) => {
   if (appDirectory) {
     localStorage.setItem('appDirectory', appDirectory);
   }
 
-  return ipcPromise.send('login-init', { userId: reqUser.username, appDirectory })
-    .then(() => {
-      const user = { ...data.user, label: reqUser.username };
+  await ipcPromise.send('login-init', { userId: reqUser.username, appDirectory });
 
-      ipcRenderer.send('login-success', data.user.id);
+  const user = { ...data.user, label: reqUser.username };
 
-      currentApiTokenKey = `${API_TOKEN_KEY}_${data.user.id}`;
+  remote.getCurrentWindow().webContents.send('login-success', data.user.id);
 
-      if (reqUser.saveLogin) {
-        localStorage.setItem(getCurrentApiTokenKey(), data.id_token);
-      } else {
-        sessionStorage.setItem(getCurrentApiTokenKey(), data.id_token);
-      }
+  ipcRenderer.on('app-init-finished', () => {
+    currentApiTokenKey = `${API_TOKEN_KEY}_${data.user.id}`;
 
-      dispatch(setUser(user));
-    });
+    if (reqUser.saveLogin) {
+      localStorage.setItem(getCurrentApiTokenKey(), data.id_token);
+    } else {
+      sessionStorage.setItem(getCurrentApiTokenKey(), data.id_token);
+    }
+
+    dispatch(setUser(user));
+  });
 };
 
 export const logout = applyAsyncLoading(() => (dispatch) => {
