### 2020-04-30 cf9e720
Completed docker manager test
