+++ packages/coinstac-docker-manager/src/index.js
@@ -31,8 +31,8 @@ const setTimeoutPromise = (delay) => {
 
 // TODO: ENV specific socket
 const docker = new Docker();
-const streamPool = {};
-const jobPool = {};
+// const streamPool = {};
+// const jobPool = {};
 let services = {};
 const portBlackList = new Set();
 
@@ -61,90 +61,90 @@ const generateServicePort = async (serviceId, start = 8100) => {
   return newPort;
 };
 
-const manageStream = (stream, jobId) => {
-  streamPool[jobId] = { stream, data: '', error: '' };
+// const manageStream = (stream, jobId) => {
+//   streamPool[jobId] = { stream, data: '', error: '' };
 
-  let header = null;
-  stream.on('readable', () => {
-    // Demux streams, docker puts stdout/err together
-    header = header || stream.read(8);
-    while (header !== null) {
-      const type = header.readUInt8(0);
-      const payload = stream.read(header.readUInt32BE(4));
-      if (payload === null) break;
-      if (type === 2) {
-        streamPool[jobId].error += payload;
-      } else {
-        streamPool[jobId].data += payload;
-      }
-      header = stream.read(8);
-    }
-  });
+//   let header = null;
+//   stream.on('readable', () => {
+//     // Demux streams, docker puts stdout/err together
+//     header = header || stream.read(8);
+//     while (header !== null) {
+//       const type = header.readUInt8(0);
+//       const payload = stream.read(header.readUInt32BE(4));
+//       if (payload === null) break;
+//       if (type === 2) {
+//         streamPool[jobId].error += payload;
+//       } else {
+//         streamPool[jobId].data += payload;
+//       }
+//       header = stream.read(8);
+//     }
+//   });
 
-  return new Promise((resolve, reject) => {
-    stream.on('end', () => {
-      const container = jobPool[jobId];
-      if (streamPool[jobId].error) {
-        container.remove()
-          .then(() => {
-            jobPool[jobId] = undefined;
-          });
-        reject(streamPool[jobId].error);
-        streamPool[jobId] = undefined;
-      } else {
-        resolve(streamPool[jobId].data);
-        streamPool[jobId] = undefined;
+//   return new Promise((resolve, reject) => {
+//     stream.on('end', () => {
+//       const container = jobPool[jobId];
+//       if (streamPool[jobId].error) {
+//         container.remove()
+//           .then(() => {
+//             jobPool[jobId] = undefined;
+//           });
+//         reject(streamPool[jobId].error);
+//         streamPool[jobId] = undefined;
+//       } else {
+//         resolve(streamPool[jobId].data);
+//         streamPool[jobId] = undefined;
 
-        container.remove()
-          .then(() => {
-            jobPool[jobId] = undefined;
-          });
-      }
-    });
-    stream.on('error', (err) => {
-      const container = jobPool[jobId];
+//         container.remove()
+//           .then(() => {
+//             jobPool[jobId] = undefined;
+//           });
+//       }
+//     });
+//     stream.on('error', (err) => {
+//       const container = jobPool[jobId];
 
-      streamPool[jobId] = undefined;
+//       streamPool[jobId] = undefined;
 
-      container.stop()
-        .then(() => container.remove())
-        .then(() => {
-          jobPool[jobId] = undefined;
-        });
-      reject(err);
-    });
-  });
-};
+//       container.stop()
+//         .then(() => container.remove())
+//         .then(() => {
+//           jobPool[jobId] = undefined;
+//         });
+//       reject(err);
+//     });
+//   });
+// };
 
-const queueJob = (jobId, input, opts) => {
-  const jobOpts = Object.assign(
-    {
-      AttachStdin: true,
[…]
