### 2020-04-30 18bd5ad
Remove user password from graphql schema
