### 2020-05-02 620c926
Fix bug on saving results
