### 2020-05-02 1765ca6
Fix promise handling on login
