### 2020-05-02 c9ad7c9
Remove console.log
