### 2020-05-02 c86c2a1
Add a delay for running pipelines
