### 2020-05-03 b0b6c31
fix err handling in sim
