+++ packages/coinstac-simulator/src/index.js
@@ -82,7 +82,8 @@ const startRun = ({
           });
       }))
         .then((errors) => {
-          if (errors.length > 0) throw errors[1];
+          // error sent to remote or the first local for local runs
+          if (errors[1] || errors[0][0]) throw errors[1] || errors[0][0];
           return {
             remote: pipelines.remote.pipeline.result,
             locals: pipelines.locals.map(local => local.pipeline.result),
