### 2020-05-04 trendscenter/coinstac#876 @yarikoptic
"Copyright (c) 2013-2018 GitHub Inc." of the downloaded releases
I have downloaded coinstac-linux-x64 for a release or linux from https://github.com/trendscenter/coinstac/releases . LICENSE shipped within it states

Copyright (c) 2013-2018 GitHub Inc.

which is likely to be incorrect
