### 2020-05-04 trendscenter/coinstac#879 @yarikoptic
Please provide LICENSE file
Would be useful for people to avoid digging through docs/README.md to find a license.  Also seems to be desired for a JOSS submission: https://github.com/openjournals/joss-reviews/issues/2166

Related: #876
