### 2020-05-05 10ce809
Updated resolver to use mongodb
