### 2020-05-05 67c26ff
Fixed showing success message issue after resetting password
