### 2020-05-05 trendscenter/coinstac#881 @yarikoptic
joss reviewer #2 comment: manuscript
I am another reviewer on JOSS submission (see https://github.com/openjournals/joss-reviews/issues/2166). In the course of the coinstac (not manuscript content) review, I have filed [other issues](https://github.com/trendscenter/coinstac/issues/created_by/yarikoptic) pertinent to the JOSS review checklist. Here I would like to provide a summary of the paper (from Mon Mar 30 10:19:56 2020 EDT, md5sum 84b7642d511894822a63769fff13525e).

First of all let me state that COINSTAC seems (have not used it personally yet to provide empirical support, but have heard many great things) indeed to be a **great** and needed to the ecosystem platform, which is a result of great work of the project's "scientific advisory board" and the developers!   Well done!  With that in mind, I feel that JOSS submission and some parts of the documentation could be improved to make it more accessible.

## Decentralized

"decentralized" should be explained and comparison to other platforms need to be adjusted:

- "decentralized": I think the reader would benefit from a brief summary of "decentralized". Moreover, the first use of the word is not really accenting on the right application per se: "can build ... **models** ... collaboratively in a decentralized fashion". COINSTAC enables decentralized computation, orchestrated via a central service. So I would argue that if anything, **models** are actually built centrally, where individual computations are decentralized.
- Picture is worth a thousand words, and I think that is the case here. Please add a figure which could help reader to understand the flow and the aspect of the decentralization of computation
- Re "the first ... platform enabling decentralized analysis": Might be worth referencing ENIGMA study as one of the early examples (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4008818/) and then guide comparison with other platforms (not tools!) while accenting on the fact that COINSTAC was **designed specifically to support** such decentralized computation, unlike those other platforms platforms (which as @gkiar points out in #849 with CBRAIN) might already support it.

# Data Protection

Wording should be adjusted to possibly avoid making false claims

- Even the statement of the need for "enhanced protection" (via differential privacy) undermines the not really described claim that COINSTAC "preserves the privacy".  Since results of computation are transferred back to the central node, and they could potentially incorporate sensitive data, there is a possibility of leaking sensitive information.  So wording should be adjusted to not "over-promise".

- Similarly fuzzy wording in the "keep your data safely ... untouched" in the  features.  "untouched" here means that data is not modified, but it is "looked at", thus there is a possibility of leakage.

- Summarizing: nothing is stated about mechanisms to protect against malicious "apps/computations" which might send "interesting" private data back. Re "untouched" - is there a guarantee that the app does not actually modify the data?  Is there any process in place for curating the "analyses/pipelines/etc" as to what "bits and pieces" would be sent to central node?

## Is it a singleton?

From my limited experience with running 'coinstac' -- there is only one central node is out there and I was not sure if it could be configurable, i.e. that I could have my own COINSTAC instance/network with coinstac client connecting to it instead of the default one.  Is COINSTAC like COINS in being "the one and only" instance?  Please clarify in the text (and/or README.md) on how to deploy/work with an independent COINSTAC deployment.

## "Conclusions"

"... central compute nodes ... attempt to draw conclusions"  sounds "too AI". Nodes are not attempting to draw conclusions I hope, but rather follow some prescribed procedures to establish the overall result based on some coded logic/stats/etc.  Please adjust wording avoiding the notion of an AI being in charge of the main node ;)

## "Example usage"

I think paper (see above recommendation about Figure) could be greatly improved with some real-world'ish example.  It would not need to be detailed, but could have references to all the details (full pipeline on github, sample data somewhere) while depicting the apps/data (as results) flow.
It then also would allow to mention that apps/pipelines are formalized in javascript (for better or for worse) which could orient a new reader better.

## Differential privacy

Could benefit from a sentence summary + reference for more details.

## "Community guidelines:"

Nothing in the paper.  At least a brief summary of what you would expect community to do (or not to do ;)) besides using (singleton?) COINSTAC, would be valuable to the reader.

## Minor typos

in the paper (as well as in other documents within repository, see #880 ) there is some minor inconsistency in opening sentences word casing.  ". we also", ". bits and pieces", ". instead" - please fixup
