### 2020-05-05 #2166 @yarikoptic
FTR: my review of the manuscript with some additional references -- https://github.com/trendscenter/coinstac/issues/881
