### 2020-05-06 trendscenter/coinstac#876 @rssk
This is the license file for Electron which our build tool is placing automatically, I was under the impression that it was using the license in the build root but that does not appear to be the case. It seems like the best solution is place our license and include but rename the Electron license as required by their license
