### 2020-05-06 trendscenter/coinstac#879 @rssk
Agreed, will address
