### 2020-05-06 2727a95
add correct LICENSES fix license build issues
