+++ packages/coinstac-ui/scripts/build.js
@@ -1,15 +1,14 @@
 const pify = require('util').promisify;
 const packager = require('electron-packager');
 const archiver = require('archiver');
-const fs = require('fs');
+const fs = require('fs').promises;
+const { createWriteStream } = require('fs');
 const rm = pify(require('rimraf'));
 const path = require('path');
 
-const rename = pify(fs.rename);
-const copy = pify(fs.copyFile);
 
 const options = {
-  all: true,
+  // all: true,
   asar: true,
   dir: `${__dirname}/../`,
   icon: path.resolve(__dirname, '../img/icons/coinstac'),
@@ -19,7 +18,7 @@ const options = {
   prune: true,
 };
 rm('./build/apps/coinstac-*')
-  .then(() => rename('./config/local.json', './config/local-build-copy.json'))
+  .then(() => fs.rename('./config/local.json', './config/local-build-copy.json'))
   .catch((e) => {
     if (e.code !== 'ENOENT') {
       throw e;
@@ -28,34 +27,39 @@ rm('./build/apps/coinstac-*')
   .then(() => {
     if (process.argv[2] && process.argv[2] === 'development') {
       console.log('Creating development build'); // eslint-disable-line no-console
-      return copy('./config/local-development.json', './config/local.json');
+      return fs.copyFile('./config/local-development.json', './config/local.json');
     } else if (process.argv[2] && process.argv[2] === 'production') { // eslint-disable-line no-else-return
       console.log('Creating production build'); // eslint-disable-line no-console
-      return copy('./config/local-production.json', './config/local.json');
+      return fs.copyFile('./config/local-production.json', './config/local.json');
     }
   })
   .then(() => packager(options))
   .then((appPaths) => {
-    appPaths.forEach((appPath) => {
-      const zip = archiver.create('zip');
-      console.log(`Finished building at: ${appPath}`); // eslint-disable-line no-console
-      console.log('Now archiving...'); // eslint-disable-line no-console
+    return Promise.all(appPaths.map(async (appPath) => {
+      await fs.rename(path.join(appPath, 'LICENSE'), path.join(appPath, 'LICENSE.electron'));
+      await fs.copyFile(path.join('.', 'LICENSE'), path.join(appPath, 'LICENSE'));
+      await new Promise((resolve, reject) => {
+        const zip = archiver.create('zip');
+        console.log(`Finished building at: ${appPath}`); // eslint-disable-line no-console
+        console.log('Now archiving...'); // eslint-disable-line no-console
 
-      const write = fs.createWriteStream(`${appPath}.zip`);
+        const write = createWriteStream(`${appPath}.zip`);
 
-      zip.pipe(write);
-      zip.on('error', (err) => {
-        throw err;
+        zip.pipe(write);
+        zip.on('error', (err) => {
+          reject(err);
+        });
+        zip.directory(appPath, false)
+          .finalize();
+
+        write.on('close', () => {
+          console.log(`Finished zipping ${appPath}.zip`); // eslint-disable-line no-console
+          resolve();
+        });
       });
-      zip.directory(appPath, false)
-        .finalize();
-
-      write.on('close', () => {
-        console.log(`Finished zipping ${appPath}.zip`); // eslint-disable-line no-console
-      });
-    });
+    }));
   })
-  .then(() => rename('./config/local-build-copy.json', './config/local.json'))
+  .then(() => fs.rename('./config/local-build-copy.json', './config/local.json'))
   .catch((err) => {
     if (err.code !== 'ENOENT') console.error('Build failed with:', err); // eslint-disable-line no-console
   });
