### 2020-05-07 8cf7d8e
Fix bug on page reload
