### 2020-05-07 25752cf
Fix bug on page reload
