### 2020-05-07 0234111
Fix subscription for all users
