+++ packages/coinstac-api-server/src/data/resolvers.js
@@ -1099,7 +1099,7 @@ const resolvers = {
     userChanged: {
       subscribe: withFilter(
         () => pubsub.asyncIterator('userChanged'),
-        (payload, variables) => (variables.userId || payload.userId === variables.userId)
+        (payload, variables) => (!variables.userId || payload.userId === variables.userId)
       )
     },
     /**
