### 2020-05-07 d0e8a2b
Updated change user app role resolver
