### 2020-05-07 5003b05
Return entities on mutation
