### 2020-05-08 f0f16f2
Updated add computation resolver
