### 2020-05-08 74b3883
fix tech typos and add up to date guide for comp dev
