### 2020-05-08 234fd2f
remove old lucidchart we dont have access to
