+++ algorithm-development/coinstac-development-guide.md
@@ -1,7 +1,7 @@
 # Coinstac Development Guide
 [![Coinstac](https://github.com/MRN-Code/coinstac/raw/master/packages/coinstac-ui/img/icons/png/64x64.png)](https://github.com/MRN-Code/coinstac)
 
-Coinstac is a Electron based application environment for decentralized algorithms, this guide aims to describe developing an alogrithm in Coinstac.
+Coinstac is a Electron based application environment for decentralized algorithms, this guide aims to describe developing an alogrithm for Coinstac.
 
 
 ### Requirements
