+++ algorithm-development/coinstac-development-guide.md
@@ -5,7 +5,7 @@ Coinstac is a Electron based application environment for decentralized algorithm
 
 
 ### Requirements
-
+  - An LTS release of [Node and NPM](https://nodejs.org/en/)
   - lastest version of [coinstac-simulator](https://npm.org/packages/coinstac-simulator)
     ```sudo npm i -g coinstac-simulator```
   - latest version of [docker](https://docs.docker.com/install/)
