### 2020-05-11 trendscenter/coinstac#849 @hvgazula
> ### General
> > > * [ ]    Repository link in the submission is out of date (redirect from `MRN-Code` currently works, but the link should be corrected since a new one is available prior to publication)
>
> > Thanks for pointing this out. I am not sure how to change the repository link in the submission. Please advise.
>
> @cMadan - please advise.
>
Still awaiting a response on how to change the repository link in the submission.
> > > * [ ]    The list of authors is inconsistent with the list of contributors for this project. Please clarify why none of the six top contributors to this project are authors, and why many of the listed authors are listed despite not having made clear contributions to the project according to the project's commit log.
>
> > Some of the top contributors (from github list) are no longer associated with the project and some are functioning as paid consultants on the project, in addition computation author contributions are in multiple other repositories. We have reworked the author list to be more comprehensive, we now include everyone who is actively participating in either the platform or computational code development or management and oversight of the project. We have also added a list of those who left the COINSTAC team prior to the start of this submission to the acknowledgement section for completeness.
>
> Thanks for your explanation and updates, @hvgazula. I'm unsure of JOSS' policy on how closely the author list should match public contributions list, if "past" contributors should be given credit, consultants, etc., but I am content with this and will defer to the editor of this article, @cMadan, to let us know if we need to discuss this further.
>
Thanks to @cMadan for the clarification.
> ### Functionality
> > > * [ ]    After following the installation instructions [here](https://github.com/trendscenter/coinstac/blob/paper-branch/SETUP.md), it appears as though I was able to get to the point of launching `coinstac-ui`, but it hangs on a loading screen in the electron app after I launch `npm start`. There is nothing in the debug/console log to suggest why this is the case.
>
> > Those instructions are for developing and contributing to the coinstac codebase. For end user instructions please check [here](https://github.com/trendscenter/coinstac-instructions). A link to these instructions has also been added to the main project README.
>
> Looking at the [raw README](https://raw.githubusercontent.com/trendscenter/coinstac/paper-branch/README.md) and searching for the URL you just shared, https://github.com/trendscenter/coinstac-instructions, it still appears to not be linked. Please fix this. Also, upon following that link myself, I see that many of the files are in the inaccessible `.docx` format. Please convert these an easily web-viewable format, such as `.md`.
>
The instructions have been converted in markdown as suggested.
> > > * [ ]    I cannot test the functionality of the app until I get past the point above
>
> > See above.
>
> I will try again once the documents are converted to a format other than `.docx` which I can view in the browser.
>
This has been taken care of.
> ### Documentation
> > > * [ ]    While the `npm` commands take care of installing dependencies, it would be nice to have the packages that are installed stated explicitly in the documentation.
>
> > The packages are listed [here](https://github.com/trendscenter/coinstac/blob/master/package.json)
>
> Thank you - please link this in `SETUP.md`.
>
A link to the list of dependencies has been placed on SETUP.md.
> > > * [ ]    While `SETUP.md` contains a link to a YouTube video showing how to use the tool, there is no accessible (text/figures only) version of these instructions for those with limited bandwidth, who I understand to be a target audience for this tool.
>
> > The instruction are [here](https://github.com/trendscenter/coinstac-instructions)
>
> Thanks. As mentioned above, can you please a) convert these to a web accessible format, and b) link them from the main repository or move them to live inside this repo directly?
>
The instructions in a web-readable format have been made available and can be found [here](https://github.com/trendscenter/coinstac-instructions). You can find a link to the instructions on the main project [README](https://github.com/trendscenter/coinstac).

> > > * [ ]    I was unable to find API-level documentation on how to use the functionality presented within the ui; while the installation did not appear to work as intended, stated above, I cannot really provide a more specific example as to what pieces are unclear or perhaps don't have integrated documentation
>
> > Please see instruction [here](https://github.com/trendscenter/coinstac-instructions/blob/master/How%20to%20Run%20an%20Analysis%20on%20COINSTAC.docx)
>
> See above.
Done.
>
> > > * [ ]    No clear automated testing/continuous integration setup
>
> > End to end integration tests can be found in the packages/coinstac-ui. We are currently in the process for rewriting necessary dependencies for use with CircleCI.
>
> Thanks, I see the folder now. Please add instructions for how to run these tests, and can you please provide an estimated timeline on how long it will take until automated testing is included or a description of why that is currently not possible?
We are currently in the process of changing over to mongodb for a variety of reasons, but one of these is for easier CI integration. We plan to have this done in the next month and CI integration soon after. We’ll add testing descriptions to the top-level MD as well.
>
> > > * [ ]    Limited contribution guidelines for developers or those seeking support
> > >
> > >   * No contributions markdown file or mention of how developers can become involved in the readme
> > >   * Issue template is good for app crashes but limited for all other forms of support
> > >   * Pull request template is incomplete
>
> > We should make a contributions template. App errors have been the most challenging for end users to submit so the template leans that way, but could be expanded. Unsure how the PR template is incomplete but open to suggestions.
>
> There are some useful contributions guide templates [here](https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors#examples-of-contribution-guidelines), for your reference. Potentially adding a "generic" issue template for non bug reports would also be useful and serves as a catch all. Apologies for being unclear; the [current pull request template](https://github.com/trendscenter/coinstac/blob/paper-branch/.github/PULL_REQUEST_TEMPLATE.md) contains necessary headings, but I intended to suggest that it could be more verbose and descriptive so contributors are aware of what information is considered "sufficient" for their contribution. There is a collection of many templates, both for issues and PRs, [here](https://github.com/stevemao/github-issue-templates).
>
We can definitely add more descriptive PR  templates, as well it appears you can have multiple templates for issues, which seems like a good solution.

> ### Software Paper
> > > * [ ]    The paper makes no mention of other tools that have similar aims, such as CBRAIN, Loris, XNAT, OpenNeuro, BrainLife, or others.
>
> > The aforementioned tools although are open-source informatics platforms dedicated to imaging-based research none of them are decentralized platforms. COINSTAC is rather a decentralized, privacy preserving analysis tool, and as such it is quite distinct in this regard from the tools above and also from even the COINS platform which is a centralized repository tool. which is exactly what COINSTAC is. We have also added references to each of the above tools, but do not do a more detailed comparison as COINSTAC is, to our knowledge, the first of its kind application platform enabling decentralized analysis of brain imaging data.
>
> Thanks for your thoughtful response, @hvgazula. I do know that CBRAIN does in fact support decentralization, Loris and XNAT put lots of care into privacy, and BrainCODE (which I forgot to link above) does both. Please include a more detailed description of what exactly COINSTAC novelly contributes relative to each of these tools, including its predecessor COINS, as that currently is a bit murky. The claim that "_COINSTAC is, to our knowledge, the first of its kind application platform enabling decentralized analysis of brain imaging data_", in my opinion, requires such a "_more detailed comparison_".
>
> > > * [ ]    Once the above-mentioned tools are explicitly stated and compared, they should be referenced accordingly.
>
> > COINSTAC is, to our knowledge, the first of its kind application platform enabling decentralized analysis of brain imaging data.
>
> See above.

COINS and LORIS are web-based data management systems that facilitate data sharing between sites whereas in COINSTAC no individual-level data is shared at all. The first BrainCODE publication was released in 2018 and does not mention support any kind of decentralized computation. Looking at CBRAIN, it was not exactly clear to me if the decentralized computation is indeed a feature of CBRAIN. Probably, the confusion could have arisen from the lack of distinction between distributed and decentralized computation. To us, distributed means: data is shared, but the computations can occur across different nodes on a cluster (for example) whereas decentralized (or federated learning) means: the data is not shared, but only the parameters during the training process that facilitate model building are shared between sites and in addition, the computations themselves can be distributed as well.
