### 2020-05-11 trendscenter/coinstac#849 @hvgazula
> @hvgazula, I've manually updated the repository link.

Thank you very much @cMadan!
