### 2020-05-13 14c364b
Refactor database credentials management on api
