+++ packages/coinstac-api-server/test/test-setup.js
@@ -1,6 +1,7 @@
 require('trace');
 require('clarify');
-const database = require('../src/database');
+const { ObjectID } = require('mongodb');
+const database = require('../db-deploy');
 const helperFunctions = require('../src/auth-helpers');
 
 const dsneMS = require('./data/coinstac-dsne-ms');
@@ -26,33 +27,33 @@ const local = require('./data/coinstac-local-test');
 const localError = require('./data/coinstac-local-error');
 
 const CONSORTIA_IDS = [
-  database.createUniqueId(),
-  database.createUniqueId(),
+  new ObjectID(),
+  new ObjectID(),
 ];
 
 const PIPELINE_IDS = [
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
 ];
 
 const COMPUTATION_IDS = [
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
-  database.createUniqueId(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
+  new ObjectID(),
 ];
 
 async function populateComputations() {
@@ -792,129 +793,142 @@ async function populateRuns() {
 }
 
 async function populateUsers() {
+  const db = database.getDbInstance();
+
   const password = await helperFunctions.hashPassword('password');
-
-  await helperFunctions.createUser({
-    username: 'test1',
-    institution: 'mrn',
-    email: 'test@mrn.org',
-    permissions: {
-      computations: {},
-      consortia: {
-        [CONSORTIA_IDS[0]]: ['member'],
-        [CONSORTIA_IDS[1]]: ['owner', 'member'],
-      },
-      pipelines: {},
-    },
-    consortiaStatuses: {
-      [CONSORTIA_IDS[0]]: 'none',
-      [CONSORTIA_IDS[1]]: 'none',
-    },
-  }, password);
-
-  await helperFunctions.createUser({
-    username: 'test2',
-    institution: 'mrn',
-    email: 'test2@mrn.org',
-    permissions: {
-      computations: {},
-      consortia: {
-        [CONSORTIA_IDS[0]]: ['member'],
-        [CONSORTIA_IDS[1]]: ['member'],
-      },
-      pipelines: {},
-    },
-    consortiaStatuses: {
-      [CONSORTIA_IDS[0]]: 'none',
-      [CONSORTIA_IDS[1]]: 'none',
-    },
-  }, password);
-
-  await helperFunctions.createUser({
-    username: 'test3',
-    institution: 'mrn',
-    email: 'test3@mrn.org',
-    permissions: {
-      computations: {},
-      consortia: {
-        [CONSORTIA_IDS[0]]: ['member'],
-        [CONSORTIA_IDS[1]]: ['member'],
-      },
-      pipelines: {},
-    },
-    consortiaStatuses: {
[…]
