### 2020-05-13 ff98fb5
Call function for db deploy
