### 2020-05-14 d9f41e3
Patching a few bugs and refactor on user update
