+++ packages/coinstac-graphql-schema/src/shared-fields.js
@@ -108,6 +108,8 @@ const sharedFields = {
     id
     email
     username
+    photo
+    photoID
   `,
   resultFields: `
     id
