### 2020-05-14 b3d1ab1
Patching a few bugs and refactor on user update
