### 2020-05-15 a76dee3
Update algorithm-onboarding.md
