### 2020-05-15 82479e3
Fixed typo in compspec
