### 2020-05-15 0e2079d
Update algorithm-onboarding.md
