### 2020-05-15 548bb45
Update algorithm-onboarding.md
Changing 'defaultValue' to 'default' in compspec
