### 2020-05-15 536fa32
Update algorithm-onboarding.md
Adding an example github repo url with naming best practice.
