### 2020-05-15 28c0ae9
Update algorithm-onboarding.md
