### 2020-05-15 4c0b98d
Update algorithm-onboarding.md
