### 2020-05-15 b8f5d1a
Update algorithm-onboarding.md
fixing path typo
