### 2020-05-15 763fe16
Update algorithm-onboarding.md
Changing 'true' to 'True' for correct python syntax
