### 2020-05-15 8cccf0d
Update algorithm-onboarding.md
