+++ algorithm-development/algorithm-onboarding.md
@@ -176,7 +176,7 @@ For now, let's just use one client and a starting value of `1`. With this starti
 To access input variables in either the `local` or `remote` sides of the computation, we use the parsed JSON input:
 ```python
 input = json.loads(sys.stdin.read())
-ourInputVariable = input["input"]["startingValue"];
+ourInputVariable = input["input"]["startingValue"]
 ```
 To send this to the remote we simply put it into the output object that is sent to stdout:
 ```python
@@ -188,14 +188,14 @@ sys.stdout.write(json.dumps(output))
 Now on the remote side of things we read things in similarly, but with the twist of the `input` document being nested for each site by the site name. We'll get how to grab the site name programmatically, but for now we know it will be `local0` as that is how site names are structured for simulator
 ```python
 inputFromLocal = json.loads(sys.stdin.read())
-inputFromLocal0 = input["input"]["local0"]["startingValue"];
+inputFromLocal0 = input["input"]["local0"]["startingValue"]
 ```
 Since all we're doing in this example is just echoing the starting value out, we finish with:
 
 ```python
-inputFromLocal0 = input["input"]["local0"]["startingValue"];
+inputFromLocal0 = input["input"]["local0"]["startingValue"]
 
-output = { "output": { outputNumber: inputFromLocal0 }, "success": True }
+output = { "output": { "outputNumber": inputFromLocal0 }, "success": True }
 sys.stdout.write(json.dumps(output))
 ```
 Note the `success: true`, this tells Coinstac that the computation has finished and the results are ready. The last output from the remote is the final output for the computation
