### 2020-05-15 07d865a
Update algorithm-onboarding.md
Fixing some stuff in iteration example code.
