### 2020-05-15 3a922d1
Add last step, some syntax changes
