+++ algorithm-development/algorithm-onboarding.md
@@ -134,7 +134,7 @@ input = json.loads(sys.stdin.read()) # input data is read in from stdin and pars
 output = { "output": {} }
 sys.stdout.write(json.dumps(output)) # output data is serialized into JSON then put onto stdout
 ```
-####### Quick note on local vs decentralized
+###### Quick note on local vs decentralized
 local and decentralized computations are made nearly the same, except for two key differences: Decentralized computations **_always_** are ended by the remote, and decentralized computations have a `"remote"` section in their compspec.
 
 ##### Creating an inputspec
@@ -200,7 +200,7 @@ sys.stdout.write(json.dumps(output))
 ```
 Note the `success: true`, this tells Coinstac that the computation has finished and the results are ready. The last output from the remote is the final output for the computation
 
-####### Running the computation
+###### Running the computation
 Now that we have our compspec, local, and remote scripts, now let's try to run it
 First we'll want to build our docker box, on some machines (typically windows) you may have to run these commands as an **administrator**. On linux you can get around this by adding your user to the `docker` group
 
@@ -326,7 +326,7 @@ Coinstac makes a trade off between messaging reliability and data transfer by us
 
 First lets put some files in the simulator test directories
 
-####### Adding test files in simulator
+###### Adding test files in simulator
 Adding files to simulator is a bit of a manual process at the moment, they must be put into the `test` folder under the automatic site name and run name the simulator generates. Sim names each site `local#` where # is 0 to the number of clients the run has. Here's what the directory structure looks like for `local0`, or the first client
 
 Let's create a file called `startingValue.txt` and `effect.txt` and place it in this directory the contents of the files will just be the variables from the `inputspec` respectively for each site, for local0 for example:
@@ -344,7 +344,8 @@ effect.txt
 ```sh
 ├── test
 │   ├── inputspec.json
-│   ├── local0 # local site name
+│   ├── input
+│   │   ├── local0 # local site name
 │   │   └── simulatorRun # the run name in all simulator pipelines, in the UI this would be a database ID
 │   │       ├── startingValue.txt
             ├── effect.txt
@@ -391,8 +392,10 @@ effectFile = open(os.path.join(input["state"]["baseDirectory"], input["input"]["
 effect = effectFile.read()
 ```
 
-##### Transferring files between clients
-Files can be transferred client to remote and remote to client by writing to a special directory. The _transfer_ directory can be access while in a computation via the `state->transferDiretory` key, looking like `inputJSON['state']['transferDiretory']` in python.
+For this exercise read both the effect and startingValue file in, replicating the multiplication operation from exercise three. Instead of sending the `withEffect` result out to the remote directly, we'll transfer it via file
+
+###### Transferring files between clients
+Files can be transferred client to remote and remote to client by writing to a special directory. The _transfer_ directory can be accessed while in a computation via the `state->transferDiretory` key, looking like `inputJSON['state']['transferDirectory']` in python.
 
 Any file put into the transfer directory is then moved from that directory to the input directory on the corresponding client when that iteration ends.
 From `remote` to `client` would look like this:
@@ -428,79 +431,40 @@ A Client:
 Remote:
 ```
 ├── ['state']['baseDiretory']
-│   ├── ['state']['baseDiretory']['clientID'] // files moved into folders based on the sending Client's ID so they cannot conflict
+│   ├── ['state']['baseDiretory']['clientID'] 
+│   │   │            // files moved into folders based on the sending Client's ID so they cannot conflict
 │   │   └── move-this-file.txt
 ```
[…]
