### 2020-05-15 d376126
Update algorithm-onboarding.md
