### 2020-05-15 d72e864
Update algorithm-onboarding.md
