### 2020-05-18 2035649
Completed client core test
