### 2020-05-19 262c3b2
Fix object notation in user arrays
