+++ packages/coinstac-ui/app/render/state/graphql/functions.js
@@ -190,7 +190,7 @@ export const UPDATE_CONSORTIA_MAPPED_USERS_MUTATION = gql`
 `;
 
 export const SAVE_MESSAGE_MUTATION = gql`
-  mutation saveMessage($threadId: ID, $title: String!, $recipients: [String!], $content: String!, $action: ActionInput)
+  mutation saveMessage($threadId: ID, $title: String!, $recipients: JSON, $content: String!, $action: ActionInput)
     ${mutations.saveMessage}
 `;
 
