### 2020-05-19 a7c806c
Fix object notation in user arrays
