+++ packages/coinstac-ui/app/render/utils/helpers.js
@@ -6,15 +6,7 @@ export const isPipelineOwner = (permissions, owningConsortium) => {
 };
 
 export const isUserInGroup = (userId, groupArr) => {
-  let res = false;
-  if(typeof groupArr === 'object'){
-    groupArr = Object.values(groupArr);
-  }
-  groupArr.map((item) => {
-    if(Object.keys(item).indexOf(userId) !== -1){
-      res = true;
-      return;
-    }
-  });
-  return res;
+  return Array.isArray(groupArr)
+    ? groupArr.findIndex(user => user === userId)
+    : userId in groupArr;
 };
