+++ packages/coinstac-api-server/src/data/resolvers.js
@@ -4,7 +4,6 @@ const GraphQLJSON = require('graphql-type-json');
 const Promise = require('bluebird');
 const { PubSub, withFilter } = require('graphql-subscriptions');
 const axios = require('axios');
-const { uniq } = require('lodash');
 const { ObjectID } = require('mongodb');
 const helperFunctions = require('../auth-helpers');
 const initSubscriptions = require('./subscriptions');
@@ -96,8 +95,8 @@ async function addUserPermissions(args) {
     };
 
     const consortiaUpdateResult = await db.collection('consortia').findOneAndUpdate({ _id: doc }, {
-      $addToSet: {
-        [`${role}s`]: { [args.userId]: userName },
+      $set: {
+        [`${role}s.${args.userId}`]: userName,
       },
     }, { returnOriginal: false });
 
@@ -147,8 +146,8 @@ async function removeUserPermissions(args) {
 
   if (args.table === 'consortia') {
     const updateObj = {
-      $pull: {
-        [`${args.role}s`]: { [args.userId]: userUpdateResult.value.username },
+      $unset: {
+        [`${args.role}s.${args.userId}`]: userUpdateResult.value.username,
       },
     };
 
@@ -334,8 +333,8 @@ const resolvers = {
 
       const runs = await db.collection('runs').find({
         $or: [
-          { clients: credentials.username },
-          { sharedUsers: credentials.username }
+          { clients: credentials.id },
+          { sharedUsers: credentials.id }
         ]
       }).toArray();
 
@@ -345,9 +344,7 @@ const resolvers = {
       const db = database.getDbInstance();
 
       const threads = await db.collection('threads').find({
-        users: {
-          $elemMatch: { username: credentials.username }
-        }
+        [`users.${credentials.id}`]: { $exists: true }
       }).toArray();
 
       return transformToClient(threads);
@@ -371,7 +368,7 @@ const resolvers = {
       await db.collection('computations').replaceOne({
         _id: args.computationSchema.id,
       }, {
-        ...args.computationSchema, submittedBy: credentials.id
+        ...args.computationSchema, submittedBy: ObjectID(credentials.id)
       }, {
         upsert: true
       });
@@ -420,9 +417,7 @@ const resolvers = {
         const consortium = await db.collection('consortia').findOne({ _id: ObjectID(consortiumId) });
         const pipeline = await fetchOnePipeline(consortium.activePipelineId);
 
-        const clientArray = consortium.members.map((client) => {
-          return Object.keys(client)[0];
-        });
+        const clientArray = Object.keys(consortium.members);
 
         const isPipelineDecentralized = pipeline.steps.findIndex(step => step.controller.type === 'decentralized') > -1;
 
@@ -551,11 +546,11 @@ const resolvers = {
       const db = database.getDbInstance();
       let consortium = await db.collection('consortia').findOne({ _id: ObjectID(args.consortiumId) });
 
-      if (consortium.members.indexOf(credentials.username) !== -1) {
+      if (credentials.id in consortium.members) {
         return consortium;
       }
 
-      await addUserPermissions({ userId: credentials.id, userName: credentials.username, role: 'member', doc: ObjectID(args.consortiumId), table: 'consortia' });
+      await addUserPermissions({ userId: ObjectID(credentials.id), userName: credentials.username, role: 'member', doc: ObjectID(args.consortiumId), table: 'consortia' });
 
       return helperFunctions.getUserDetails(credentials.username);
     },
@@ -567,7 +562,7 @@ const resolvers = {
      * @return {object} Updated consortium
      */
     leaveConsortium: async ({ auth: { credentials } }, args) => {
-      await removeUserPermissions({ userId: credentials.id, role: 'member', doc: ObjectID(args.consortiumId), table: 'consortia' });
[…]
