### 2020-05-28 29c6e3b
Added coinstac api server test
