+++ packages/coinstac-api-server/seed/test-setup.js
@@ -1,0 +1,3 @@
+const populate = require('./populate');
+
+populate();
