+++ packages/coinstac-api-server/src/routes/auth.js
@@ -44,19 +44,19 @@ module.exports = [
       pre: [
         { method: helperFunctions.validateUniqueUser },
       ],
-      handler: (req, res) => {
-        helperFunctions.hashPassword(req.payload.password)
-          .then(passwordHash => helperFunctions.createUser(req.payload, passwordHash))
-          .then(({
-            _id, username, institution, email, permissions,
-          }) => {
-            res({
-              id_token: helperFunctions.createToken(username),
-              user: {
-                id: _id, username, institution, email, permissions,
-              },
-            }).code(201);
-          });
+      handler: async (req, res) => {
+        const passwordHash = await helperFunctions.hashPassword(req.payload.password);
+        const user = await helperFunctions.createUser(req.payload, passwordHash);
+        const {
+          _id, username, institution, email, permissions,
+        } = user;
+
+        res({
+          id_token: helperFunctions.createToken(username),
+          user: {
+            id: _id, username, institution, email, permissions,
+          },
+        }).code(201);
       },
     },
   },
