+++ algorithm-development/computations.md
@@ -1,0 +1,80 @@
+# Computations
+[![Coinstac](https://github.com/MRN-Code/coinstac/raw/master/packages/coinstac-ui/img/icons/png/64x64.png)](https://github.com/MRN-Code/coinstac)
+
+Below is a list of COINSTAC computations that have been implemented.
+
+## Currently Available
+
+- Regression on Freesurfer Volumes
+  - Creator: [Harsh Gazula](https://github.com/hvgazula)
+  - Maintainer: [Harsh Gazula](https://github.com/hvgazula)
+  - Repository: https://github.com/trendscenter/coinstac-regression-fsl
+  
+- Regression on VBM Maps
+  - Creator: [Harsh Gazula](https://github.com/hvgazula)
+  - Maintainer: [Harsh Gazula](https://github.com/hvgazula)
+  - Repository: https://github.com/trendscenter/coinstac-regression-vbm
+  
+- Single-shot Ridge Regression on VBM Maps
+  - Creator: [Harsh Gazula](https://github.com/hvgazula)
+  - Maintainer: [Harsh Gazula](https://github.com/hvgazula)
+  - Repository: https://github.com/trendscenter/coinstac-ssr-vbm
+  
+- Multi-shot Ridge Regression on VBM Maps
+  - Creator: [Harsh Gazula](https://github.com/hvgazula)
+  - Maintainer: [Harsh Gazula](https://github.com/hvgazula)
+  - Repository: https://github.com/trendscenter/coinstac-msr-vbm
+  
+- Multi-shot Ridge Regression on Freesurfer Volumes
+  - Creator: [Harsh Gazula](https://github.com/hvgazula)
+  - Maintainer: [Harsh Gazula](https://github.com/hvgazula)
+  - Repository: https://github.com/trendscenter/coinstac-msr-fsl
+
+- ENIGMA SANS
+  - Creator: [Harsh Gazula](https://github.com/hvgazula)
+  - Maintainer: [Ross Kelly](https://github.com/rssk)
+  - Repository: https://github.com/trendscenter/coinstac-enigma-sans
+
+- ENIGMA
+  - Creator: [Harsh Gazula](https://github.com/hvgazula)
+  - Maintainer: [Harsh Gazula](https://github.com/hvgazula)
+  - Repository: https://github.com/trendscenter/coinstac-enigma
+  
+- Spatially Constrained ICA
+  - Creator: [Bradley Baker](https://github.com/bbradt)
+  - Maintainer: [Bradley Baker](https://github.com/bbradt)
+  - Repository: https://github.com/trendscenter/coinstac_spatially_constrained_ica
+  
+- Distributed Dynamic Functional Network Connectivity Pipeline
+  - Creator: [Bradley Baker](https://github.com/bbradt)
+  - Maintainer: [Bradley Baker](https://github.com/bbradt)
+  -Repository: https://github.com/trendscenter/coinstac_ddfnc_pipeline
+
+- Distributed t-Stochastic Neighbor Embedding
+  - Creator: [Debratta Saha](https://github.com/deblearn)
+  - Maintainer: [Debratta Saha](https://github.com/deblearn)
+  - Repository: https://github.com/trendscenter/coinstac-dsne-multishot/tree/master/dsne-ms_test
+
+## Compatible with Simulator, but not yet in UI
+
+- Decentralized MANCOVA
+  - Creator: [Bradley Baker](https://github.com/bbradt)
+  - Maintainer: [Bradley Baker](https://github.com/bbradt)
+  - Repository: https://github.com/trendscenter/coinstac-mancova
+  
[…]
