### 2020-06-04 71eeedb
Refactor pipeline/maps on freesurfer
