+++ packages/coinstac-api-server/test/test-setup.js
@@ -20,6 +20,7 @@ const fmri = require('./data/coinstac-fmri');
 
 const decentralized = require('./data/coinstac-decentralized-test');
 const transfer = require('./data/coinstac-file-transfer-test');
+const stress = require('./data/coinstac-file-stress-test');
 const decentralizedError = require('./data/coinstac-decentralized-error');
 const enigmaSans = require('./data/coinstac-enigma-sans');
 const local = require('./data/coinstac-local-test');
@@ -60,6 +61,7 @@ helperFunctions.getRethinkConnection()
         Object.assign({}, local, { submittedBy: 'author' }),
         Object.assign({}, decentralized, { submittedBy: 'author' }),
         Object.assign({}, transfer, { submittedBy: 'author' }),
+        Object.assign({}, stress, { submittedBy: 'author' }),
         // Object.assign({}, ssrFsl, { submittedBy: 'author' }),
         Object.assign({}, msrFsl, { submittedBy: 'author' }),
         Object.assign({}, msrVbm, { submittedBy: 'author' }),
