### 2020-06-08 e90a749
Remove references of 'data' and 'covariates'
