### 2020-06-08 e93c8ce
Remove references of 'data' and 'covariates'
