### 2020-06-08 2e9f7c4
Remove references of 'data' and 'covariates'
