+++ packages/coinstac-ui/app/main/utils/run-pipeline-functions.js
@@ -1,193 +1,20 @@
-const { resolve, extname, sep } = require('path');
-
-function stepInputNeedsDataMapping(inputSchema) {
-  return 'ownerMappings' in inputSchema;
-}
-
-function stepInputMapSchemaHasOptions(inputMapSchemaKeys) {
-  return inputMapSchemaKeys.includes('data') && inputMapSchemaKeys.findIndex(element => element.includes('options'))
-    && !inputMapSchemaKeys.includes('covariates') && !inputMapSchemaKeys.includes('file');
-}
-
-function parseValue(value, type) {
-  const val = value.toLowerCase();
-
-  switch (type) {
-    case 'boolean':
-      return val === 'true';
-    case 'number':
-      return parseFloat(val);
-    default:
-      return null;
-  }
-}
-
-function inputSchemaHasFileSource(inputSchemaObj) {
-  return inputSchemaObj.source === 'file';
-}
-
-function encodeFilePath(filePath) {
-  const escape = (string) => {
-    return string.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&'); // eslint-disable-line no-useless-escape
-  };
-  const pathsep = new RegExp(`${escape(sep)}|:`, 'g');
-
-  if (filePath.includes(sep)) {
-    return filePath.replace(pathsep, '-');
-  }
-
-  return null;
-}
-
-function makeDataObjFromDataMappings(dataMappings, stepIndex) {
-  const dataObj = {};
-  const dataMaps = dataMappings.dataMappings[stepIndex].data;
-  dataMaps.map((map) => {
-    const key = map.pipelineVariableName;
-    dataObj[key] = encodeFilePath(map.dataFileFieldName);
-  });
-  return dataObj;
-}
-
-function mapVariablesIntoArray(inputSchema, dataMappings, data, baseDirectory) {
-  const dataArray = [];
-  const variablesArray = [];
-  const variablesTypesArray = [];
-
-  const variablesTypes = {};
-
-  inputSchema.ownerMappings.forEach((variable) => {
-    const mapping = dataMappings.find(m => m.pipelineVariableName === variable.name);
-    variablesTypes[mapping.dataFileFieldName] = variable.type;
-  });
-
-  const headerRow = Object.keys(data[0])
-    .sort((a, b) => {
-      if (!(a in variablesTypes)) {
-        return -1;
-      }
-
-      if (!(b in variablesTypes)) {
-        return 1;
-      }
-
-      return a < b ? -1 : 1;
-    });
-
-  dataArray.push(headerRow);
-
-  data.forEach((dataRow) => {
-    const row = [];
-
-    headerRow.forEach((headerColumn) => {
-      let value = dataRow[headerColumn];
-
-      if (headerColumn in variablesTypes) {
-        value = parseValue(value, variablesTypes[headerColumn]);
-      } else {
-        const escape = (string) => {
-          return string.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&'); // eslint-disable-line no-useless-escape
-        };
-        const pathsep = new RegExp(`${escape(sep)}|:`, 'g');
-        if (extname(value) !== '') {
-          value = resolve(baseDirectory, value).replace(pathsep, '-');
-        }
-      }
-
-      row.push(value);
-    });
-
-    dataArray.push(row);
-  });
-
-  headerRow.forEach((headerColumn) => {
-    if (headerColumn in variablesTypes) {
-      variablesArray.push(headerColumn);
-      variablesTypesArray.push(variablesTypes[headerColumn]);
-    }
-  });
-
-  return [
-    [dataArray],
-    variablesArray,
-    variablesTypesArray,
-  ];
-}
-
 function parsePipelineInput(pipeline, dataMappings) {
   const steps = [];
[…]
