### 2020-06-11 8eca670
Updated auth helpers according to change request
