+++ packages/coinstac-pipeline/src/pipeline-manager.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const mqtt = require('mqtt');
 const { promisify } = require('util');
 const mkdirp = promisify(require('mkdirp'));
-const rimraf = promisify(require('rimraf'));
+const rimraf = require('rmfr');
 const path = require('path');
 const http = require('http');
 const https = require('https');
@@ -15,9 +15,11 @@ const Emitter = require('events');
 const winston = require('winston');
 const express = require('express');
 const multer = require('multer');
-const decompress = require('decompress');
 const uuid = require('uuid/v4');
 const dockerManager = require('coinstac-docker-manager');
+const tar = require('tar-fs');
+const zlib = require('zlib');
+const merge2 = require('merge2');
 
 const readdir = promisify(fs.readdir);
 
@@ -31,6 +33,74 @@ const defaultLogger = winston.loggers.get('pipeline');
 defaultLogger.level = process.LOGLEVEL ? process.LOGLEVEL : 'info';
 
 const Pipeline = require('./pipeline');
+
+/**
+ * [splitFilesFromStream description]
+ * @param  {[type]} stream    [description]
+ * @param  {[type]} filePath  [description]
+ * @param  {[type]} chunkSize [description]
+ * @return {[type]}           [description]
+ */
+const splitFilesFromStream = (stream, filePath, chunkSize) => {
+  let currentChunk = 0;
+  let currentChunkLen = 0;
+  const asyncStreams = [];
+  const splits = [];
+  const newFileStream = fs.createWriteStream(
+    `${filePath}.${currentChunk}`
+  );
+  splits[currentChunk] = {
+    filePath: `${filePath}.${currentChunk}`,
+    stream: newFileStream,
+  };
+  asyncStreams.push(new Promise((resolve, reject) => {
+    splits[currentChunk].stream.on('close', () => resolve());
+    splits[currentChunk].stream.on('error', e => reject(e));
+  }));
+  stream.on('data', (data) => {
+    if (currentChunkLen >= chunkSize) {
+      splits[currentChunk].stream.end();
+      currentChunk += 1;
+      const newFileStream = fs.createWriteStream(
+        `${filePath}.${currentChunk}`
+      );
+      splits[currentChunk] = {
+        filePath: `${filePath}.${currentChunk}`,
+        stream: newFileStream,
+      };
+      asyncStreams.push(new Promise((resolve, reject) => {
+        splits[currentChunk].stream.on('close', () => resolve());
+        splits[currentChunk].stream.on('error', () => reject());
+      }));
+      splits[currentChunk].stream.write(data);
+      currentChunkLen = data.length;
+    } else {
+      splits[currentChunk].stream.write(data);
+      currentChunkLen += data.length;
+    }
+  });
+  asyncStreams.push(new Promise((resolve, reject) => {
+    stream.on('end', () => {
+      splits[currentChunk].stream.end();
+      resolve();
+    });
+    stream.on('error', (e) => {
+      reject(e);
+    });
+  }));
+  return Promise.all(asyncStreams)
+    .then(() => splits.map(split => path.basename(split.filePath)));
+};
+
+const extractTar = (parts, outdir) => {
+  return new Promise((resolve, reject) => {
+    const unpack = merge2(...parts.map(part => fs.createReadStream(part)))
+      .pipe(zlib.createGunzip())
+      .pipe(tar.extract(outdir));
+    unpack.on('finish', () => resolve());
+    unpack.on('error', e => reject(e));
+  });
+};
 
 /**
  * get files in a dir recursively
@@ -175,10 +245,11 @@ module.exports = {
         memo.push((async () => {
           let retryLimit = 0;
           let success = false;
+          let partFile;
           // retry 1000 times w/ backout
           while (retryLimit < limit) {
             try {
[…]
