### 2020-06-18 f3c6082
Refactor maps and compspec for dsne and enigma
