### 2020-06-18 766d6e5
Refactor maps and compspec for dsne and enigma
