### 2020-06-18 ef0788d
Updated README to direct users to setup page
