### 2020-06-18 852ad85
Added remote url to auth redux store
