### 2020-06-18 9aef498
Added remote url to auth redux store
