### 2020-06-18 3558935
Passed remote url to client core
