+++ packages/coinstac-ui/app/main/utils/boot/configure-core.js
@@ -4,7 +4,7 @@ const CoinstacClientCore = require('coinstac-client-core');
 const { merge } = require('lodash');
 const parseCLIInput = require('./parse-cli-input.js');
 
-module.exports = function configureCore(config, logger, userId, appDirectory) {
+module.exports = function configureCore(config, logger, userId, appDirectory, remoteURL) {
   const coreConfiguration = merge(
     JSON.parse(config.toString()),
     parseCLIInput.get(),
@@ -12,6 +12,7 @@ module.exports = function configureCore(config, logger, userId, appDirectory) {
       logger,
       userId,
       appDirectory,
+      remoteURL,
     }
   );
 
