+++ packages/coinstac-ui/app/config.js
@@ -56,6 +56,7 @@ const conf = convict({
     win32: 'coinstac/',
   },
   coinstacHome: path.join(home(), '.coinstac'),
+  remoteURL: path.join(home(), '.coinstac'),
 });
 
 module.exports = function loadConfig() {
