### 2020-06-18 a87395a
add npm i step to dev builds
