+++ packages/coinstac-ui/scripts/build-setup.js
@@ -3,17 +3,19 @@ const rmrf = pify(require('rimraf'));
 const mkdirp = pify(require('mkdirp'));
 const ncp = pify(require('ncp'));
 const path = require('path');
+const npmi = require('./utils/npmi');
 
+const paths = [
+  path.join('node_modules', 'coinstac-common'),
+  path.join('node_modules', 'coinstac-client-core'),
+  path.join('node_modules', 'coinstac-graphql-schema'),
+  path.join('node_modules', 'coinstac-docker-manager'),
+  path.join('node_modules', 'coinstac-pipeline'),
+];
 rmrf('./node_modules/coinstac-*(common|client-core|graphql-schema|docker-manager|pipeline)')
   .then(() => rmrf('./node_modules/.bin/coinstac-*(common|graphql-schema|docker-manager|pipeline)'))
   .then(() => rmrf('./config/local.json'))
-  .then(() => Promise.all([
-    mkdirp(path.join('node_modules', 'coinstac-common')),
-    mkdirp(path.join('node_modules', 'coinstac-client-core')),
-    mkdirp(path.join('node_modules', 'coinstac-graphql-schema')),
-    mkdirp(path.join('node_modules', 'coinstac-docker-manager')),
-    mkdirp(path.join('node_modules', 'coinstac-pipeline')),
-  ]))
+  .then(() => Promise.all(paths.map(p => mkdirp(p))))
   .then(() => Promise.all([
     ncp('../coinstac-common/package.json', './node_modules/coinstac-common/package.json'),
     ncp('../coinstac-common/src', './node_modules/coinstac-common/src'),
@@ -27,4 +29,5 @@ rmrf('./node_modules/coinstac-*(common|client-core|graphql-schema|docker-manager
     ncp('../coinstac-pipeline/package.json', './node_modules/coinstac-pipeline/package.json'),
     ncp('../coinstac-pipeline/src', './node_modules/coinstac-pipeline/src'),
   ]))
+  .then(() => paths.map(p => npmi(p)))
   .catch(console.error); // eslint-disable-line no-console
