### 2020-06-18 5244f73
fix build process for local and dev finalfinal.pdf
