### 2020-06-19 04a4f8f
retry err logging
