### 2020-06-19 a2301be
more file logging
