### 2020-06-19 7c609dd
fix file retry logging
