### 2020-06-23 a0bbe15
tune resend and chunk size
