### 2020-06-24 #2166 @hvgazula
Hello @kthyng Thanks for checking. I shall be pasting my comments sometime tonight.
