### 2020-06-25 1d4911c
Save client server url in client core
