+++ packages/coinstac-ui/app/main/utils/boot/configure-core.js
@@ -4,7 +4,7 @@ const CoinstacClientCore = require('coinstac-client-core');
 const { merge } = require('lodash');
 const parseCLIInput = require('./parse-cli-input.js');
 
-module.exports = function configureCore(config, logger, userId, appDirectory, remoteURL) {
+module.exports = function configureCore(config, logger, userId, appDirectory, clientServerURL) {
   const coreConfiguration = merge(
     JSON.parse(config.toString()),
     parseCLIInput.get(),
@@ -12,7 +12,7 @@ module.exports = function configureCore(config, logger, userId, appDirectory, re
       logger,
       userId,
       appDirectory,
-      remoteURL,
+      clientServerURL,
     }
   );
 
