### 2020-06-25 ed8e815
Created coinstac client server
