+++ packages/coinstac-client-server/src/routes/pipeline.js
@@ -1,0 +1,13 @@
+/* eslint-disable max-len */
+
+module.exports = [
+  {
+    method: 'GET',
+    path: '/',
+    config: {
+      handler: (_, res) => {
+        res({}).code(200);
+      },
+    },
+  },
+];
