+++ packages/coinstac-client-server/src/routes/index.js
@@ -1,0 +1,4 @@
+const pipeline = require('./pipeline');
+
+// Add additional route files below
+module.exports = pipeline;
