### 2020-06-25 7cae9f8
Added removing multiple docker images to docker manager
