+++ packages/coinstac-docker-manager/src/index.js
@@ -618,6 +618,9 @@ const removeImage = (imageId) => {
   return docker.getImage(imageId).remove();
 };
 
+const removeImagesFromList = comps => Promise.all(comps.map(image => removeImage(`${image}:latest`)))
+  .then(streams => streams.map((stream, index) => ({ stream, compId: comps[index] })));
+
 /**
  * Attempts to stop a given service
  * If there are no other users, the service stops
@@ -675,6 +678,7 @@ module.exports = {
   pullImagesFromList,
   pruneImages,
   removeImage,
+  removeImagesFromList,
   queueJob,
   setLogger,
   startService,
