### 2020-06-25 ec3917b
Added pull and remove docker images to the client server
