+++ packages/coinstac-client-server/src/index.js
@@ -3,7 +3,6 @@ const hapi = require('hapi');
 process.LOGLEVEL = 'silly';
 
 const config = require('./config');
-const routes = require('./routes');
 
 const server = new hapi.Server();
 server.connection({
@@ -11,8 +10,10 @@ server.connection({
   port: config.hapiPort,
 });
 
-server.route(routes);
-server.start((startErr) => {
-  if (startErr) throw startErr;
-  console.log(`Server running at: ${server.info.uri}`); // eslint-disable-line no-console
+require('./routes').then((routes) => {
+  server.route(routes);
+  server.start((startErr) => {
+    if (startErr) throw startErr;
+    console.log(`Server running at: ${server.info.uri}`); // eslint-disable-line no-console
+  });
 });
