### 2020-06-25 1e2d3c4
Added stopping pipeline route
