### 2020-06-28 trendscenter/coinstac#881 @hvgazula
> I am another reviewer on JOSS submission (see [openjournals/joss-reviews#2166](https://github.com/openjournals/joss-reviews/issues/2166)). In the course of the coinstac (not manuscript content) review, I have filed [other issues](https://github.com/trendscenter/coinstac/issues/created_by/yarikoptic) pertinent to the JOSS review checklist. Here I would like to provide a summary of the paper (from Mon Mar 30 10:19:56 2020 EDT, md5sum 84b7642d511894822a63769fff13525e).
>
> First of all let me state that COINSTAC seems (have not used it personally yet to provide empirical support, but have heard many great things) indeed to be a **great** and needed to the ecosystem platform, which is a result of great work of the project's "scientific advisory board" and the developers! Well done! With that in mind, I feel that JOSS submission and some parts of the documentation could be improved to make it more accessible.
>
> ## Decentralized
> "decentralized" should be explained and comparison to other platforms need to be adjusted:
>
> * "decentralized": I think the reader would benefit from a brief summary of "decentralized". Moreover, the first use of the word is not really accenting on the right application per se: "can build ... **models** ... collaboratively in a decentralized fashion". COINSTAC enables decentralized computation, orchestrated via a central service. So I would argue that if anything, **models** are actually built centrally, where individual computations are decentralized.
Thank you very much for the comment. I added ‘aka decentralization’ in the text. We don’t disagree with your interpretation but at the same time, we reckon, the sentence ‘can build...models...collaboratively in a decentralized fashion‘ can also be interpreted as the decentralized computations collaborating to build a model. Think of a team of players playing a game. The result of the game (here model) is not decentralized but a result of the individuals playing (decentralized computations).

> * Picture is worth a thousand words, and I think that is the case here. Please add a figure which could help reader to understand the flow and the aspect of the decentralization of computation
We concur with you about the suitability of a pictorial representation of the decentralization of computation but unfortunately such a representation is tied to the specific algorithm and might not necessarily do complete justice. However, we added a figure to depict the overall structure of the coinstac framework with local computer nodes (where the computations happen) and a central node (where the model is built).

> * Re "the first ... platform enabling decentralized analysis": Might be worth referencing ENIGMA study as one of the early examples (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4008818/) and then guide comparison with other platforms (not tools!) while accenting on the fact that COINSTAC was **designed specifically to support** such decentralized computation, unlike those other platforms platforms (which as @gkiar points out in #849 with CBRAIN) might already support it.
Thank you very much for pointing out the need for comparing with existing frameworks. We’ve added a ‘Related work’ section where ENIGMA, ViPAR are discussed. We would also like to bring to your attention that in your understanding CBRAIN does not necessarily support decentralization but supports distributed computation. The distinction has been clarified in a response to @gkiar.
>
> # Data Protection
> Wording should be adjusted to possibly avoid making false claims
>
> * Even the statement of the need for "enhanced protection" (via differential privacy) undermines the not really described claim that COINSTAC "preserves the privacy".  Since results of computation are transferred back to the central node, and they could potentially incorporate sensitive data, there is a possibility of leaking sensitive information.  So wording should be adjusted to not "over-promise".
Thanks for pointing it out. The sentence has been rephrased as follows: Because there is no pooling of data it can technically guarantee the privacy of individual datasets. However, we also offer differentially private algorithms for enhanced protection.

> * Similarly fuzzy wording in the "keep your data safely ... untouched" in the  features.  "untouched" here means that data is not modified, but it is "looked at", thus there is a possibility of leakage.
Modified to ‘technically untouched’‘. However, the following point ‘Applying differential privacy strategies to further enhance anonymization of private data, whilst still permitting collaboration.’ should clarify it further.

> * Summarizing: nothing is stated about mechanisms to protect against malicious "apps/computations" which might send "interesting" private data back. Re "untouched" - is there a guarantee that the app does not actually modify the data?  Is there any process in place for curating the "analyses/pipelines/etc" as to what "bits and pieces" would be sent to central node?
Please see the bullet point at the bottom of the paper under (3).

>
> ## Is it a singleton?
> From my limited experience with running 'coinstac' -- there is only one central node is out there and I was not sure if it could be configurable, i.e. that I could have my own COINSTAC instance/network with coinstac client connecting to it instead of the default one. Is COINSTAC like COINS in being "the one and only" instance? Please clarify in the text (and/or README.md) on how to deploy/work with an independent COINSTAC deployment.
Anyone can set up their own COINSTAC server. We already provide instructions at https://github.com/trendscenter/coinstac/blob/master/SETUP.md, and have linked to them on the main README.md page.

>
> ## "Conclusions"
> "... central compute nodes ... attempt to draw conclusions" sounds "too AI". Nodes are not attempting to draw conclusions I hope, but rather follow some prescribed procedures to establish the overall result based on some coded logic/stats/etc. Please adjust wording avoiding the notion of an AI being in charge of the main node ;)
Changed to “...central compute nodes … help the user draw conclusions”

>
> ## "Example usage"
> I think paper (see above recommendation about Figure) could be greatly improved with some real-world'ish example. It would not need to be detailed, but could have references to all the details (full pipeline on github, sample data somewhere) while depicting the apps/data (as results) flow.
> It then also would allow to mention that apps/pipelines are formalized in javascript (for better or for worse) which could orient a new reader better.
Please see the section entitled Example usage.

>
> ## Differential privacy
> Could benefit from a sentence summary + reference for more details.
Added a citation describing differential privacy
>
> ## "Community guidelines:"
> Nothing in the paper. At least a brief summary of what you would expect community to do (or not to do ;)) besides using (singleton?) COINSTAC, would be valuable to the reader.
Please see the last section entitled Community Guidelines.
>
> ## Minor typos
> in the paper (as well as in other documents within repository, see #880 ) there is some minor inconsistency in opening sentences word casing. ". we also", ". bits and pieces", ". instead" - please fixup
The paper has been edited for grammar.
