### 2020-06-28 trendscenter/coinstac#881 @hvgazula
@yarikoptic Thanks for reviewing this work. I've added responses to your comments. I hope you will be able to take some time off your schedule to take a look and we will be happy to address any further concerns you may have. Thanks again for your time.
