### 2020-06-29 f2d5e41
Start pipeline using client server
