### 2020-06-29 9cf9ff5
Updated client server routes on latest hapi version
