### 2020-07-01 984554e
Remove docker/ws stuff to help in debug
