### 2020-07-01 79247d2
Remove docker/ws stuff to help in debug
