### 2020-07-01 114e1b0
Fix bug on first login with new users
