+++ packages/coinstac-ui/test/e2e/computation-run-2-members.test.js
@@ -24,14 +24,12 @@ chai.use(chaiAsPromised);
 
 const app1 = new Application({
   path: electronPath,
-  env: { NODE_ENV: 'test', TEST_INSTANCE: 'test-1' },
   args: [appPath, '-r', mocksPath],
   port: 9515,
 });
 
 const app2 = new Application({
   path: electronPath,
-  env: { NODE_ENV: 'test', TEST_INSTANCE: 'test-2' },
   args: [appPath, '-r', mocksPath],
   port: 9516,
 });
@@ -101,7 +99,7 @@ describe('e2e run computation with 2 members', () => {
       .click('button=Save')
       .waitForVisible('span=Consortium Saved', EXIST_TIMEOUT)
       .click('a=Consortia')
-      .waitForVisible(`h1=${CONS_NAME}`)
+      .waitForVisible(`h5=${CONS_NAME}`)
   ));
 
   it('accesses the Add Pipeline page', () => (
