### 2020-07-03 27ff657
Fix some requires on e2e tests
