### 2020-07-04 5647065
Integrated hapi nes for socket communication
