### 2020-07-06 #2166 @hvgazula
> Hello @kthyng Thanks for checking. I shall be pasting my comments sometime tonight.

Hi @kthyng and @cMadan FYI, I submitted my responses to the second reviewer as well last week and will not wait to hear from both the reviewers if they have any comments. Thanks!
