### 2020-07-06 2b2a92e
Created client specific pipeline managers
