+++ packages/coinstac-client-server/src/index.js
@@ -1,23 +1,11 @@
 const hapi = require('@hapi/hapi');
 const nes = require('nes');
-const path = require('path');
-const PipelineManager = require('coinstac-pipeline');
 const config = require('./config');
 const routes = require('./routes');
 
 process.LOGLEVEL = 'silly';
 
 const init = async () => {
-  const manager = await PipelineManager.create({
-    mode: 'remote',
-    clientId: 'remote',
-    operatingDirectory: path.resolve(config.operatingDirectory, 'coinstac'),
-    remotePort: 3400,
-    mqttRemotePort: config.mqttServer.port,
-    mqttRemoteProtocol: config.mqttServer.protocol,
-    mqttRemoteURL: config.mqttServer.hostname,
-  });
-
   const server = hapi.Server({
     host: config.host,
     port: config.hapiPort,
@@ -25,7 +13,7 @@ const init = async () => {
 
   await server.register(nes);
 
-  server.route(routes(manager, server));
+  server.route(routes(server));
 
   await server.start();
 
