### 2020-07-07 8d1f686
Added api token validation in client server
