### 2020-07-07 c4bd76f
Passed api token from ui to client server in core
