+++ packages/coinstac-ui/app/main/utils/boot/configure-core.js
@@ -4,7 +4,9 @@ const CoinstacClientCore = require('coinstac-client-core');
 const { merge } = require('lodash');
 const parseCLIInput = require('./parse-cli-input.js');
 
-module.exports = function configureCore(config, logger, userId, appDirectory, clientServerURL) {
+module.exports = function configureCore(
+  config, logger, userId, appDirectory, clientServerURL, token
+) {
   const coreConfiguration = merge(
     JSON.parse(config.toString()),
     parseCLIInput.get(),
@@ -13,6 +15,7 @@ module.exports = function configureCore(config, logger, userId, appDirectory, cl
       userId,
       appDirectory,
       clientServerURL,
+      token,
     }
   );
 
