### 2020-07-08 01c013a
Fix mapped status and add online status check
