+++ packages/coinstac-ui/app/render/state/graphql/functions.js
@@ -180,12 +180,12 @@ export const UPDATE_USER_CONSORTIUM_STATUS_MUTATION = gql`
 `;
 
 export const UPDATE_CONSORTIUM_MAPPED_USERS_MUTATION = gql`
-  mutation updateConsortiumMappedUsers($consortiumId: ID, $mappedForRun: [ID])
+  mutation updateConsortiumMappedUsers($consortiumId: ID, $isMapped: Boolean)
     ${mutations.updateConsortiumMappedUsers}
 `;
 
 export const UPDATE_CONSORTIA_MAPPED_USERS_MUTATION = gql`
-  mutation updateConsortiaMappedUsers($consortia: [ID])
+  mutation updateConsortiaMappedUsers($consortia: [ID], $isMapped: Boolean)
     ${mutations.updateConsortiaMappedUsers}
 `;
 
@@ -197,4 +197,16 @@ export const SAVE_MESSAGE_MUTATION = gql`
 export const SET_READ_MESSAGE_MUTATION = gql`
   mutation setReadMessage($threadId: ID!, $userId: ID!)
     ${mutations.setReadMessage}
+`;
+
+export const FETCH_USERS_ONLINE_STATUS = gql`
+  query fetchUsersOnlineStatus {
+    fetchUsersOnlineStatus
+  }
+`;
+
+export const USERS_ONLINE_STATUS_CHANGED_SUBSCRIPTION = gql`
+  subscription usersOnlineStatusChanged {
+    usersOnlineStatusChanged
+  }
 `;
