### 2020-07-14 a98bd35
Stop remote pipeline when user stops locally
