### 2020-07-15 497af8b
modify redic timeout
