### 2020-07-15 eac2943
pr fixes
