+++ packages/coinstac-simulator/src/index.js
@@ -17,18 +17,18 @@ const exitHook = require('exit-hook');
 const startRun = ({
   spec, runMode = 'local', clientCount = 1, operatingDirectory = 'test',
 }) => {
-  // return new Promise((resolve, reject) => {
-  //   // the execArgv opt are a work around for https://github.com/nodejs/node/issues/9435
-  //   const mqtt = fork(path.resolve(__dirname, 'mqtt-server.js'), { execArgv: [], stdio: ['inherit', 'inherit', 'inherit', 'ipc'] });
-  //   mqtt.on('message', (m) => {
-  //     if (m.e) return reject(m.e);
-  //     if (m.started) resolve();
-  //   });
-  //   exitHook(() => {
-  //     mqtt.kill();
-  //   });
-  // })
-    return Promise.resolve().then(async () => {
+  return new Promise((resolve, reject) => {
+    // the execArgv opt are a work around for https://github.com/nodejs/node/issues/9435
+    const mqtt = fork(path.resolve(__dirname, 'mqtt-server.js'), { execArgv: [], stdio: ['inherit', 'inherit', 'inherit', 'ipc'] });
+    mqtt.on('message', (m) => {
+      if (m.e) return reject(m.e);
+      if (m.started) resolve();
+    });
+    exitHook(() => {
+      mqtt.kill();
+    });
+  })
+    .then(async () => {
       const pipelines = {
         locals: [],
       };
@@ -60,9 +60,6 @@ const startRun = ({
           clientId: `local${i}`,
           mode: 'local',
           operatingDirectory: path.resolve(operatingDirectory),
-          // mqttRemoteURL: 'coinstac.rs.gsu.edu',
-          // mqttRemotePort: '80',
-          // mqttRemoteProtocol: 'mqtts:',
         });
         pipelines.locals.push({
           manager: localPipelineManager,
