+++ packages/coinstac-images/coinstac-python-base-stream/test/check.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const getStdin = require('get-stdin');
-
-getStdin().then((str) => {
-  process.stdout.write(str);
-  process.stderr.write(str);
-});

[…]
