+++ packages/coinstac-images/coinstac-python-base-stream/server-http.js
@@ -1,310 +1,0 @@
-'use strict';
-
-const net = require('net');
-require('trace');
-require('clarify');
-const { Transform } = require('stream');
-const { spawn } = require('child_process');
-const winston = require('winston');
-
-const logger = winston.createLogger({
-  level: 'info',
-  transports: [
-    new winston.transports.Console({ format: winston.format.cli() }),
-  ],
-});
-process.on('uncaughtException', (err) => {
-  logger.error(`Server Error: ${err.stack}`);
-});
-
-const createWebServer = (requestHandler) => {
-  const server = net.createServer()
-    .on('error', (err) => {
-      logger.error(`Server Error: ${err.stack}`);
-    });
-
-  const handleConnection = (socket) => {
-    // Subscribe to the readable event once so we can start calling .read()
-    socket.on('error', (e) => {
-      console.log('Server Socket error:'); // eslint-disable-line no-console
-      console.log(e); // eslint-disable-line no-console
-    });
-    socket.once('readable', () => {
-      logger.debug('Received incoming connection');
-      // Parse headers out of socket
-      let reqBuffer = Buffer.alloc(0);
-      const parseHeaders = (chunk) => {
-        reqBuffer = Buffer.concat([reqBuffer, chunk]);
-        const marker = reqBuffer.indexOf('\r\n\r\n');
-        if (marker !== -1) {
-          // If we reached \r\n\r\n, there could be data after it. Take note.
-          const remaining = reqBuffer.slice(marker + 4);
-          // The header is everything we read, up to and not including \r\n\r\n
-          const reqHeader = reqBuffer.slice(0, marker).toString();
-          // This pushes the extra data we read back to the socket's readable stream
-          socket.pause();
-          socket.unshift(remaining);
-          /* Request-related business */
-          // Start parsing the header
-          const reqHeaders = reqHeader.split('\r\n');
-          // First line is special
-          const reqLine = reqHeaders.shift().split(' ');
-          // Further lines are one header per line, build an object out of it.
-          const headers = reqHeaders.reduce((acc, currentHeader) => {
-            const [key, value] = currentHeader.split(':');
-            return {
-              ...acc,
-              [key.trim().toLowerCase()]: value.trim(),
-            };
-          }, {});
-          logger.debug(`Parsed header: ${JSON.stringify(headers)}`);
-          // This object will be sent to the handleRequest callback.
-          const request = {
-            method: reqLine[0],
-            url: reqLine[1],
-            httpVersion: reqLine[2].split('/')[1],
-            headers,
-            socket,
-          };
-          let status = 200;
-          let statusText = 'OK';
-          let headersSent = false;
-          let isChunked = false;
-
-          const responseHeaders = {
-            server: 'coinstac',
-          };
-
-          const setHeader = (key, value) => {
-            responseHeaders[key.toLowerCase()] = value;
-          };
-
-          const sendHeaders = () => {
-            // Only do this once :)
-            if (!headersSent) {
-              headersSent = true;
-              // Add the date header
-              setHeader('date', new Date().toGMTString());
-              // Send the status line
-              socket.write(`HTTP/1.1 ${status} ${statusText}\r\n`);
-              logger.info(`Response header: HTTP/1.1 ${status} ${statusText}\r\n`);
-              // Send each following header
-              Object.keys(responseHeaders).forEach((headerKey) => {
-                socket.write(`${headerKey}: ${responseHeaders[headerKey]}\r\n`);
-              });
-              // Add the final \r\n that delimits the response headers from body
-              socket.write('\r\n');
-            }
-          };
-
-          const response = {
[…]
