### 2020-07-21 c9c74f4
Use electron-builder on packaging
