### 2020-07-22 c9704c5
Test windows packaging and add mac config
