### 2020-07-22 b5d1bf0
add script to get secrets for env config
