### 2020-07-22 6964abb
Some changes on build script
