### 2020-07-27 db1fdc0
Add some files to npm ignore
