### 2020-07-30 0fc1b76
modify api & server to use envars, modify config script to suite
