+++ scripts/aws-secret-fetch.sh
@@ -1,7 +1,12 @@
-if [ $1 -eq 0 ]; then
-    echo "No MFA token provided"
-    echo "The first argument for this script should be your MFA"
-    echo "device's current valid rotating token"
+if [ $1 -eq 0 ] || [ $2 = "" ] || [ $3 = "" ]; then
+    echo "Incorrect arguments"
+    echo ""
+    echo "First argument for this script should be your MFA"
+    echo "  device's current valid rotating token"
+    echo ""
+    echo "Second argument is the config env [prod, dev, local]"
+    echo ""
+    echo "Third is the output type [export, systemd, json]"
     exit 1
 fi
 unset AWS_ACCESS_KEY_ID
@@ -16,14 +21,19 @@ SESSIONTOKEN=$(echo $SESSION | jq -r .Credentials.SessionToken)
 export AWS_ACCESS_KEY_ID=$ACCESSKEYID
 export AWS_SECRET_ACCESS_KEY=$SECRETACCESSKEY
 export AWS_SESSION_TOKEN=$SESSIONTOKEN
-CONFIG=$(aws secretsmanager get-secret-value --secret-id coinstac --query SecretString --output text)
+if [ $2 != prod ]; then
+  SECRETID="coinstac-$2"
+else
+  SECRETID="coinstac"
+fi
+CONFIG=$(aws secretsmanager get-secret-value --secret-id $SECRETID --query SecretString --output text)
 
 
 unset AWS_ACCESS_KEY_ID
 unset AWS_SECRET_ACCESS_KEY
 unset AWS_SESSION_TOKEN
 
-if [ $2 = export ]; then
+if [ $3 = export ]; then
   echo "export CLOUDINARY_UPLOAD_PRESET=$(echo $CONFIG | jq -r .CLOUDINARY_UPLOAD_PRESET)"
   echo "export CLOUDINARY_API_KEY=$(echo $CONFIG | jq -r .CLOUDINARY_API_KEY)"
   echo "export CLOUDINARY_API_SECRET=$(echo $CONFIG | jq -r .CLOUDINARY_API_SECRET)"
@@ -31,6 +41,36 @@ if [ $2 = export ]; then
   echo "export CLOUDINARY_DELETE_URL=$(echo $CONFIG | jq -r .CLOUDINARY_DELETE_URL)"
   echo "export GITHUB_ACCESS_TOKEN=$(echo $CONFIG | jq -r .GITHUB_ACCESS_TOKEN)"
   echo "export SENDGRID_API_KEY=$(echo $CONFIG | jq -r .SENDGRID_API_KEY)"
+  echo "export API_SERVER_PORT=$(echo $CONFIG | jq -r .API_SERVER_PORT)"
+  echo "export API_SERVER_HOSTNAME=$(echo $CONFIG | jq -r .API_SERVER_HOSTNAME)"
+  echo "export MONGODB_CONN_STRING=$(echo $CONFIG | jq -r .MONGODB_CONN_STRING)"
+  echo "export DATABASE_NAME=$(echo $CONFIG | jq -r .DATABASE_NAME)"
+  echo "export PIPELINE_SERVER_PORT=$(echo $CONFIG | jq -r .PIPELINE_SERVER_PORT)"
[…]
