### 2020-07-30 1a1424d
modify api & server to use envars, modify config script to suite
