### 2020-07-30 dc4854d
add env replacement for dbmap
