### 2020-07-30 7ad647f
add env replacement for dbmap
