### 2020-08-03 4d7f666
add new err handling back in
