+++ packages/coinstac-images/coinstac-file-transfer-test/Dockerfile
@@ -1,4 +1,4 @@
-FROM cb
+FROM coinstacteam/coinstac-python-base-stream
 
 # Set the working directory to /app
 WORKDIR /computation
