### 2020-08-07 3eb2bf4
jk remove binding, fix some logic errs, make reset ux smoother
