### 2020-08-10 d262694
some UI changes and add new config opts
