### 2020-08-10 8cb6265
some UI changes and add new config opts
