+++ packages/coinstac-pipeline/src/computation.js
@@ -1,4 +1,5 @@
 'use strict';
+
 const _ = require('lodash');
 
 module.exports = {
