### 2020-08-17 trendscenter/coinstac#849 @gkiar
@hvgazula please update the issue template and PR templates, ensuring your new ones match the description provided by JOSS:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I will test the functionality shortly, but otherwise I have updated my ticked off most items on my checklist given your modifications.
