### 2020-08-17 #2166 @gkiar
I have updated my review. The only remaining piece from my perspective is the addition of issue guidelines and PR templates for contributors and users, fitting the description provided in the review template above.
