### 2020-09-02 #2166 @yarikoptic
I am satisfied with edits to the paper, and have nothing you add to @gkiar 's comments.
