### 2020-09-02 #2166 @hvgazula
> I am satisfied with edits to the paper, and have nothing you add to @gkiar 's comments.

Thank you very much, @yarikoptic! With the templates too in place as @gkiar suggested, probably it's now up to @cMadan to decide on the next steps.
