### 2020-09-04 #2166 @cMadan
@hvgazula, can you fix the bib so that the capitalisation of named programs is correct in the reference list? (E.g., ENIGMA, MCIC, VBM, etc.)
