### 2020-09-04 #2166 @yarikoptic
Enclose within `{}`, so `{ENIGMA}` which will retain capitalization
