### 2020-09-04 #2166 @hvgazula
> Enclose within `{}`, so `{ENIGMA}` which will retain capitalization

Worked like a charm. Thank you very much for the suggestion, @yarikoptic.
