### 2020-09-04 #2166 @hvgazula
> > @hvgazula, can you fix the bib so that the capitalisation of specific names are correct in the reference list? (E.g., ENIGMA, MCIC, VBM, etc.)
>
> Thank you @cMadan. You are right. Indeed, the capitalizations are missing ENIGMA, MCIC and VBM in the reference list. It seems that if a title starts with all upper case it works fine. However, an all uppercase in the middle of the title is not retained in the final output. Please advise how I can address this.

@cMadan The capitalization has been taken care of.
