### 2020-09-06 #2166 @cMadan
@hvgazula, can you also add DOIs for all bib entries? See the recent comment from whedon for suggestions (but please double check these yourself!)
