### 2020-10-21 #2166 @hvgazula
https://github.com/trendscenter/coinstac/releases/tag/v5.1.3

> @hvgazula, everything looks good to me!
>
> To move forward with accepting your submission, there are a few last things to take care of:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
5.1.3 https://github.com/trendscenter/coinstac/releases/tag/v5.1.3
> * [ ]  Archive the reviewed software in Zenodo
Done
> * [ ]  Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
Done
> * [ ]  List the Zenodo DOI of the archived version here.
10.5281/zenodo.4111006
>
> You may find this helpful: https://guides.github.com/activities/citable-code/
