### 2020-10-25 #2166 @hvgazula
> Hi @hvgazula, I'm the EIC on duty this week, doing some final checks before publishing your article.
>
> The only thing I noticed is that your article is missing an explicit [Statement of Need section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), which we have begun requiring for all articles. Can you add this? You might be able to just rename an existing section, or move some of your text into the new one.

Done. Thanks for letting me know about it. Really appreciate it.
